% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{add_cache_block}
\alias{add_cache_block}
\title{Add the block id to a parse table}
\usage{
add_cache_block(pd_nested)
}
\description{
Must be after \code{\link[=nest_parse_data]{nest_parse_data()}} because requires a nested parse table as
input
}
\keyword{internal}
