% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{add_id_and_short}
\alias{add_id_and_short}
\title{Add column \code{pos_id} and \code{short}}
\usage{
add_id_and_short(pd)
}
\arguments{
\item{pd}{A flat parse table}
}
\description{
Adds column \code{pos_id} and \code{short} to a flat parse table.
}
\keyword{internal}
