% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.R
\name{add_line_col_to_wrapped_expr}
\alias{add_line_col_to_wrapped_expr}
\title{Adds line and col information to an expression from its child}
\usage{
add_line_col_to_wrapped_expr(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
Adds line and col information to an expression from its child
}
