% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-strings.R
\name{add_spaces_or_newlines}
\alias{add_spaces_or_newlines}
\alias{add_newlines}
\alias{add_spaces}
\title{Concentrate newlines or spaces in a string}
\usage{
add_newlines(n)

add_spaces(n)
}
\arguments{
\item{n}{Scalar indicating how many characters should be concentrated}
}
\value{
A string.
}
\description{
Concentrate newlines or spaces in a string
}
\keyword{internal}
