% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylerignore.R
\name{add_stylerignore}
\alias{add_stylerignore}
\title{Adds the stylerignore column}
\usage{
add_stylerignore(parse_data)
}
\description{
If a token falls within a stylerignore tag, the column is set to \code{TRUE},
otherwise to \code{FALSE}.
}
