% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-replacement.R
\name{add_to_element_above}
\alias{add_to_element_above}
\title{Add elements in a vector to the elment before it}
\usage{
add_to_element_above(x, index)
}
\arguments{
\item{x}{A vector}

\item{index}{A scalar indicating from which elements in the vector should
be added to its left neighbour.}
}
\description{
Add elements in a vector to the elment before it
}
\examples{
add_to_element_above(c(1, 2, 3), 2)
}
