% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-alignment-utils.R
\name{alignment_col1_is_named}
\alias{alignment_col1_is_named}
\title{Checks if all arguments of column 1 are named}
\usage{
alignment_col1_is_named(relevant_pd_by_line)
}
\arguments{
\item{relevant_pd_by_line}{A list with parse tables of a multi-line call,
excluding first and last column.}
}
\description{
Checks if all arguments of column 1 are named
}
\keyword{internal}
