% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-alignment-utils.R
\name{alignment_drop_comments}
\alias{alignment_drop_comments}
\title{Remove all comment tokens}
\usage{
alignment_drop_comments(pd_by_line)
}
\arguments{
\item{pd_by_line}{A list, each element corresponding to a potentially
incomplete parse table that represents all token from one line.}
}
\description{
Must be after split by line because it invalidates (lag)newlines, which are
used for splitting by line.
}
\keyword{internal}
