% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-alignment-utils.R
\name{alignment_ensure_no_closing_brace}
\alias{alignment_ensure_no_closing_brace}
\title{Ensure the closing brace of the call is removed}
\usage{
alignment_ensure_no_closing_brace(pd_by_line, last_line_droped_early)
}
\arguments{
\item{pd_by_line}{A list, each element corresponding to a potentially
incomplete parse table that represents all token from one line.}
}
\description{
Must be after dropping comments because the closing brace is only guaranteed
to be the last token in that case.
}
\keyword{internal}
