% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-alignment-utils.R
\name{alignment_ensure_trailing_comma}
\alias{alignment_ensure_trailing_comma}
\title{Ensure last pd has a trailing comma}
\usage{
alignment_ensure_trailing_comma(pd_by_line)
}
\arguments{
\item{pd_by_line}{A list, each element corresponding to a potentially
incomplete parse table that represents all token from one line.}
}
\description{
Must be after \code{\link[=alignment_ensure_no_closing_brace]{alignment_ensure_no_closing_brace()}} because if it comes after
\code{\link[=alignment_ensure_trailing_comma]{alignment_ensure_trailing_comma()}}, the last expression would not be a
brace, which would make removal complicated.
}
\keyword{internal}
