% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-alignment-utils.R
\name{alignment_has_correct_spacing_around_eq_sub}
\alias{alignment_has_correct_spacing_around_eq_sub}
\title{Check if spacing around \code{=} is correct}
\usage{
alignment_has_correct_spacing_around_eq_sub(pd_sub)
}
\arguments{
\item{pd_sub}{The subset of a parse table corresponding to one line.}
}
\description{
At least one space around \code{EQ_SUB}
}
\keyword{internal}
