% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-alignment-utils.R
\name{alignment_serialize}
\alias{alignment_serialize}
\title{Serialize text from a parse table}
\usage{
alignment_serialize(pd_sub)
}
\description{
Line breaks are ignored as they are expected to be checked in
\code{\link[=token_is_on_aligned_line]{token_is_on_aligned_line()}}.
}
\keyword{internal}
