% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-alignment-utils.R
\name{alignment_serialize_column}
\alias{alignment_serialize_column}
\title{Serialize all lines for a given column}
\usage{
alignment_serialize_column(relevant_pd_by_line, column)
}
\arguments{
\item{relevant_pd_by_line}{A list with parse tables of a multi-line call,
excluding first and last column.}

\item{column}{The index of the column to serialize.}
}
\description{
Serialize all lines for a given column
}
\keyword{internal}
