% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-alignment-utils.R
\name{alignment_serialize_line}
\alias{alignment_serialize_line}
\title{Serialize one line for a column}
\usage{
alignment_serialize_line(relevant_pd_by_line, column)
}
\arguments{
\item{relevant_pd_by_line}{A list with parse tables of a multi-line call,
excluding first and last column.}

\item{column}{The index of the column to serialize.}
}
\description{
Serialize one line for a column
}
