% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindent.R
\name{apply_ref_indention}
\alias{apply_ref_indention}
\title{Apply reference indention to tokens}
\usage{
apply_ref_indention(flattened_pd)
}
\arguments{
\item{flattened_pd}{A flattened parse table}
}
\description{
Applies the reference indention created with functions
\code{\link[=update_indention_ref]{update_indention_ref()}} to the flattened parse table. The indention
is applied to all token that inherit from a reference token sequentially,
i.e. by looping over the target tokens.
}
\keyword{internal}
