% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reindent.R
\name{apply_ref_indention_one}
\alias{apply_ref_indention_one}
\title{Applying reference indention of a target token}
\usage{
apply_ref_indention_one(flattened_pd, target_token)
}
\arguments{
\item{flattened_pd}{A flattened parse table}

\item{target_token}{The index of the token from wich the indention level
should be applied to other tokens.}
}
\description{
Applies the indention level of \code{target_token} to all tokens that have
\code{target_token} as a reference. This includes adding spaces to the first
tokens on a line and updating the column \code{col1} and \code{col2} for all token
on that line so they are kept updated.
}
