% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{apply_shift_to_tokens}
\alias{apply_shift_to_tokens}
\title{Shift the position of tokens by a certain amount of spaces}
\usage{
apply_shift_to_tokens(flattened_pd, spaces_to_update, cols_to_update, shift)
}
\arguments{
\item{spaces_to_update}{The indices of the first tokens on every line in the
parse table which should be shifted.}

\item{cols_to_update}{The indices of the tokens in the parse table which
should be shifted.}
}
\description{
Shifts tokens in a flattened parse table. For
the indices \code{spaces_to_update}, this entails increasing both the column
\code{spaces} and \code{col} by \code{shift}. For the indices \code{col_to_update}, only the
\code{col} attribute will be updated since we do not want to add extra spaces
between the tokens on the same line.
}
