% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylerignore.R
\name{apply_stylerignore}
\alias{apply_stylerignore}
\title{Ensure correct positional information for stylerignore expressions}
\usage{
apply_stylerignore(flattened_pd)
}
\description{
\itemize{
\item Get the positional information for tokens with a stylerignore tag from
\code{env_current}, which recorded that information from the input text.
\item Replace the computed lag_newlines and lag_spaces information in the parse
table with this information.
TODO also take token, as rules-replacement may have modified the tokens too.
}
}
