% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{apply_transformers}
\alias{apply_transformers}
\title{Apply transformers to a parse table}
\usage{
apply_transformers(pd_nested, transformers)
}
\arguments{
\item{pd_nested}{A nested parse table.}

\item{transformers}{A list of \emph{named} transformer functions}
}
\description{
Depending on whether \code{transformers} contains functions to modify the
line break information, the column \code{multi_line} is updated (after
the line break information is modified) and
the rest of the transformers is applied afterwards, or (if line break
information is not to be modified), all transformers are applied in one
step. The former requires two pre visits and one post visit, the latter
only one pre visit.
}
