% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialized_tests.R
\name{as_is}
\alias{as_is}
\title{Pass input right through}
\usage{
as_is(x)
}
\arguments{
\item{x}{The an R oject}
}
\description{
Passes an input as is without changing its class (unlike \code{\link[base:I]{base::I()}}.
Helpful for testing in conjunction with \code{\link[utf8:transform_lines_enc]{utf8::transform_lines_enc()}}.
}
