% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-assert-args.R
\name{assert_filetype}
\alias{assert_filetype}
\title{Make sure all supplied file types are allowed}
\usage{
assert_filetype(lowercase_filetype)
}
\arguments{
\item{lowercase_filetype}{A vector with file types to check, all lower case.}
}
\description{
Make sure all supplied file types are allowed
}
