% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communicate.R
\name{assert_R.cache_installation}
\alias{assert_R.cache_installation}
\title{Assert the R.cache installation in conjunction with the cache config}
\usage{
assert_R.cache_installation(installation_only = FALSE, action = "abort")
}
\arguments{
\item{installation_only}{Whether or not to only check if R.cache is
installed.}
}
\description{
R.cache needs to be installed if caching functionality is enabled
}
\keyword{internal}
