% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-assert-args.R
\name{assert_text}
\alias{assert_text}
\title{Assert text to be of positive length and replace it with the empty
string otherwise.}
\usage{
assert_text(text)
}
\arguments{
\item{text}{The input to style.}
}
\description{
Assert text to be of positive length and replace it with the empty
string otherwise.
}
