% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-assert-args.R
\name{assert_tokens}
\alias{assert_tokens}
\title{Check token validity}
\usage{
assert_tokens(tokens)
}
\arguments{
\item{tokens}{Tokens to check.}
}
\description{
Check whether one or more tokens exist and have a unique token-text mapping
}
