% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-assert-args.R
\name{assert_transformers}
\alias{assert_transformers}
\title{Assert the transformers}
\usage{
assert_transformers(transformers)
}
\arguments{
\item{transformers}{A list of transformer functions that operate on flat
parse tables.}
}
\description{
Actually only assert name and version of style guide in order to make sure
caching works correctly.
}
\keyword{internal}
