% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{autoupdate}
\alias{autoupdate}
\title{Auto-update your hooks}
\usage{
autoupdate(root = here::here())
}
\arguments{
\item{root}{The path to the root directory of your project.}
}
\description{
Runs \verb{pre-commit autoupdate}.
}
