% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.R
\name{bind_with_child}
\alias{bind_with_child}
\title{Bind a parse table with one of its children}
\usage{
bind_with_child(pd_nested, pos)
}
\arguments{
\item{pd_nested}{A nested parse table.}

\item{pos}{The position of the child to bind.}
}
\description{
Bind a parse table with one of its children and return parse table, ordered
according to the appearance of the tokens.
}
