% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{cache_activate}
\alias{cache_activate}
\alias{cache_deactivate}
\title{Activate or deactivate the styler cache}
\usage{
cache_activate(cache_name = NULL)

cache_deactivate(verbose = TRUE)
}
\arguments{
\item{cache_name}{The name of the cache to use. If
\code{NULL}, the option "styler.cache_name" is considered which defaults to
the version of styler used.}
}
\description{
Helper functions to control the behavior of caching. Simple wrappers around
\code{\link[base:options]{base::options()}}.
}
\seealso{
Other cache managers: \code{\link{cache_clear}},
  \code{\link{cache_info}}
}
\concept{cache managers}
