% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-block.R
\name{cache_find_block}
\alias{cache_find_block}
\title{Every expression is an expression itself, but consecutive comments are one
expression (potentially)}
\usage{
cache_find_block(pd)
}
\description{
Every expression is an expression itself, but consecutive comments are one
expression (potentially)
}
