% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{cache_info}
\alias{cache_info}
\title{Show information about the styler cache}
\usage{
cache_info(cache_name = NULL, format = "lucid")
}
\arguments{
\item{cache_name}{The name of the cache to use. If
\code{NULL}, the option "styler.cache_name" is considered which defaults to
the version of styler used.}

\item{format}{Either "lucid" for a printed summary or "tabular" for a
tabular summary from \code{\link[base:file.info]{base::file.info()}}.}
}
\description{
If the cache is used at all or not is determined via the R option
\code{styler.use_cache}.
}
\seealso{
Other cache managers: \code{\link{cache_activate}},
  \code{\link{cache_clear}}
}
\concept{cache managers}
