% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{cache_is_activated}
\alias{cache_is_activated}
\title{Check if a cache is activated}
\usage{
cache_is_activated(cache_name = NULL)
}
\arguments{
\item{cache_name}{The name of the cache to check. If \code{NULL}, we check if
any cache is activated. If not \code{NULL}, we check if a specific cache is
activated.}
}
\description{
Check if a cache is activated
}
\keyword{internal}
