% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{cache_more_specs}
\alias{cache_more_specs}
\title{Create more specs}
\usage{
cache_more_specs(include_roxygen_examples, base_indention)
}
\description{
Syntactic sugar for creating more specs. This is useful when we want to add
more arguments (because we can search for this function in the source code).
}
\keyword{internal}
