% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{cache_write}
\alias{cache_write}
\title{Write to the cache}
\usage{
cache_write(text, transformers, more_specs)
}
\arguments{
\item{text}{Code to create a cache for. This should be styled text, as the
approach used by \{styler\} does not cache input, but styled code.}

\item{transformers}{A list of transformer functions, because we can only
know if text is already correct if we know which transformer function it
should be styled with.}
}
\description{
Write to the cache
}
\keyword{internal}
