% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{call_and_capture}
\alias{call_and_capture}
\title{Make a call with \code{\link[=system2]{system2()}} and capture the effects.}
\usage{
call_and_capture(command, args, ...)
}
\arguments{
\item{command}{The command to issue. A character string of length one.}

\item{args}{The command line arguments.}

\item{...}{Arguments passed to \code{\link[=system2]{system2()}}.}
}
\value{
A list with:
\itemize{
\item content of stderr
\item content of stdout
\item exit status
}
}
\description{
Make a call with \code{\link[=system2]{system2()}} and capture the effects.
}
\keyword{internal}
