% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calls_sys}
\alias{calls_sys}
\title{Invoke a system command}
\usage{
calls_sys(sys_call, ...)
}
\arguments{
\item{sys_call}{The call to be executed.}

\item{...}{Arguments passed to \code{\link[=shell]{shell()}} or \code{\link[=system]{system()}}.}
}
\description{
Wraps a system command into \code{\link[=shell]{shell()}} or \code{\link[=system]{system()}}, depending on the
operating system.
}
\keyword{internal}
