% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-files.R
\name{can_verify_roundtrip}
\alias{can_verify_roundtrip}
\title{Check whether a round trip verification can be carried out}
\usage{
can_verify_roundtrip(transformers)
}
\arguments{
\item{transformers}{The list of transformer functions used for styling.
Needed for reverse engineering the scope.}
}
\description{
If scope was set to "line_breaks" or lower (compare \code{\link[=tidyverse_style]{tidyverse_style()}}),
we can compare the expression before and after styling and return an error if
it is not the same.
}
\keyword{internal}
