% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing-public-api.R
\name{catch_style_file_output}
\alias{catch_style_file_output}
\title{Capture and post-process the output of \code{style_file} without causing side
effects}
\usage{
catch_style_file_output(file_in = c("public-api", "xyzdir-dirty",
  "dirty-sample-with-scope-tokens.R"), encoding)
}
\arguments{
\item{file_in}{A vector passed to \code{\link[=testthat_file]{testthat_file()}} to construct the path
to the reference file.}
}
\value{
A character vector with the captured output of \code{\link[=style_file]{style_file()}} called on
\code{file_in} ran in a temp dir to avoid side effects on the input file (because
the next time the test would ran, the file would not need styling).
In this output, the
path has been removed so only the file name is contained in the return
value to make the output portable across systems (in particular, to run it on
CI systems). Since the horizontal rules width depend on the length of the
path, the rules were also removed to standardize the ouptut of this function.
}
\description{
Capture and post-process the output of \code{style_file} without causing side
effects
}
\keyword{internal}
