% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_guides.R
\name{character_to_ordered}
\alias{character_to_ordered}
\title{Convert a character vector to an ordered factor}
\usage{
character_to_ordered(x, levels, name = substitute(x))
}
\arguments{
\item{x}{A character vector.}

\item{levels}{A vector with levels.}

\item{name}{The name of the character vector to be displayed if the
construction of the factor fails.}
}
\description{
Convert a vector to an ordered factor but stop if any of the values in
\code{x} does not match the predefined levels in \code{levels.}
}
\keyword{internal}
