% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_pd.R
\name{child_indents}
\alias{child_indents}
\title{Check whether a child will indent}
\usage{
child_indents(pd, opening, token)
}
\arguments{
\item{pd}{A parse table.}

\item{opening}{The row number of the opening token in the parse table.}

\item{token}{On which token the indention check should be based on.}
}
\value{
Returns \code{TRUE} if \code{pd} has at least one child that indents starting
on the same line as the opening token, \code{FALSE} otherwise.
}
\description{
Check whether a child will indent
}
