% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{child_will_indent}
\alias{child_will_indent}
\title{Check whether a child will re-indent token-dependent}
\usage{
child_will_indent(flattened_pd, target_index, cols_to_update)
}
\arguments{
\item{flattened_pd}{A flattened parse table.}

\item{target_index}{The index of the token in \code{flattened_pd} that has the
indention we want to apply to other tokens.}

\item{cols_to_update}{The indices of the tokens in the parse table which
should be shifted.}
}
\value{
\code{TRUE} if there is a child in \code{flattened_pd} that will be
re-indented because of some other token that causes token-dependent
indention, \code{FALSE} if that is not the case.
}
\description{
Check whether a child in a flattened parse table will be re-indented
depending on another token.
}
