% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{communicate_captured_call}
\alias{communicate_captured_call}
\title{Communicate a captured call}
\usage{
communicate_captured_call(x, preamble = "")
}
\arguments{
\item{x}{The output of \code{\link[=call_and_capture]{call_and_capture()}}.}
}
\description{
Communicate a captured call
}
\keyword{internal}
