% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communicate.R
\name{communicate_summary}
\alias{communicate_summary}
\title{Communicate the summary of styling}
\usage{
communicate_summary(changed, ruler_width)
}
\arguments{
\item{changed}{Boolean with indicating for each file whether or not it has
been changed.}

\item{ruler_width}{Integer used to determine the width of the ruler.}
}
\description{
Communicate the summary of styling
}
\keyword{internal}
