% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communicate.R
\name{communicate_warning}
\alias{communicate_warning}
\title{Communicate a warning if necessary}
\usage{
communicate_warning(changed, transformers)
}
\arguments{
\item{changed}{Boolean with indicating for each file whether or not it has
been changed.}

\item{transformers}{The list of transformer functions used for styling.
Needed for reverse engineering the scope.}
}
\description{
If round trip verification was not possible, issue a warning to review the
changes carefully.
}
\keyword{internal}
