% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{complete_styled_expr}
\alias{complete_styled_expr}
\title{Complete styled expression with unstyled fraction on start / end line}
\usage{
complete_styled_expr(context, styled_expr, neighbourhood)
}
\arguments{
\item{context}{The context from \code{\link[=find_active_context]{find_active_context()}}.}

\item{styled_expr}{Character vector consisting of the styled expression.}

\item{neighbourhood}{Neighbourhood obtained from \code{\link[=extract_neighbourhood]{extract_neighbourhood()}}.}
}
\description{
Essentially adding the parts before the highlighted area on the line where
the higlighted area starts to the styled expression. Also, after the styled
expression, the remainder of that line that was not styled is added.
}
