% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{compute_cols_to_update}
\alias{compute_cols_to_update}
\title{Return the indices of the tokens for which \code{col} needs to be udpated}
\usage{
compute_cols_to_update(flattened_pd, target_index)
}
\arguments{
\item{flattened_pd}{A flattened parse table.}

\item{target_index}{The index of the token in \code{flattened_pd} that has the
indention we want to apply to other tokens.}
}
\description{
\code{col} needs to be updated for all tokens that lay on a line which will be
re-indented. This is equivalent to be positioned between the opening
parenthesis (given by \code{target_index}) and closing parenthesis (which is to
be found from \code{target_index} via the \code{id} attribute) in \code{flattened_pd}.
}
\seealso{
compute_spaces_to_update
}
