% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_pd.R
\name{compute_indent_indices}
\alias{compute_indent_indices}
\title{Compute the indices that need indention}
\usage{
compute_indent_indices(pd, token = "'('", indent_last = FALSE)
}
\arguments{
\item{pd}{A parse table.}

\item{token}{A character vector with tokens.}

\item{indent_last}{Flag to indicate whether the last token in \code{pd} should
be indented or not. See 'Details'.}
}
\description{
Based on \code{token}, find the rows in \code{pd} that need to be indented.
}
\details{
For example when \code{token} is a parenthesis, the closing parenthesis does not
need indention, but if token is something else, for example a plus (+), the
last token in \code{pd} needs indention.
}
