% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{compute_parent_to_update}
\alias{compute_parent_to_update}
\title{Compute the indices of the parents that are to be updated}
\usage{
compute_parent_to_update(flattened_pd)
}
\arguments{
\item{flattened_pd}{A flattened parse table.}
}
\description{
Compute the indices of the parents that are to be updated
}
