% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsed.R
\name{compute_parse_data_flat_with_ws}
\alias{compute_parse_data_flat_with_ws}
\title{Parse and pre-process character vector}
\usage{
compute_parse_data_flat_with_ws(text)
}
\arguments{
\item{text}{A character vector.}
}
\value{
A pre-processed parse table.
}
\description{
The function obtains detailed parse information for \code{text} via
\code{\link[utils:getParseData]{utils::getParseData()}} and does some minimal pre-processing by calling
\code{\link[=enhance_parse_data]{enhance_parse_data()}}.
}
\details{
Roughly speaking, this is the inverse operation of
\code{\link[=serialize_parse_data_flat]{serialize_parse_data_flat()}}, which turns a parse table into a character
vector, since \code{compute_parse_data_flat_with_ws()} turns a character vector
into a parse table.
}
\section{Details}{

Preprocessing includes
\itemize{
\item removing non-terminal entries.
\item removing columns id, parent and terminal.
\item adding a start token.
\item adding line-break and space information.
\item removing spaces in comments at the end of the line.
}
}

