% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{compute_shift_from_col}
\alias{compute_shift_from_col}
\title{How much should we shift?}
\usage{
compute_shift_from_col(flattened_pd, target_index, subject_index)
}
\arguments{
\item{flattened_pd}{A flattened parse table.}

\item{target_index}{The index of the token in \code{flattened_pd} that has the
indention we want to apply to other tokens.}

\item{subject_index}{The index of a token in \code{flattened_pd} that
should match the target indention after re-indention.}
}
\description{
Compute the amount of spaces needed to be inserted before some tokens to
match the target indention.
}
