% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{compute_spaces_to_update}
\alias{compute_spaces_to_update}
\title{Which spaces need an update?}
\usage{
compute_spaces_to_update(flattened_pd, cols_to_update)
}
\arguments{
\item{flattened_pd}{A flattened parse table.}

\item{cols_to_update}{The indices of the tokens in the parse table which
should be shifted.}
}
\description{
Given indices for which cols needs to be updated, find spaces that need to
be updated. Only the tokens in \code{cols_to_update} that lay on a new line need
their \code{spaces} attribute to be updated. All tokens on the same line don't
need extra re-indention, since they are going to be shifted by the same
amount as the first token.
}
\seealso{
compute_cols_to_update
}
