% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{compute_start_from_target}
\alias{compute_start_from_target}
\title{Compute the start given a target index}
\usage{
compute_start_from_target(flattened_pd, target_index)
}
\description{
Computes the index of the first token for which \code{col} needs to be updated.
This is the first token on the line following \code{target_index}.
}
