% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{compute_stop_from_start}
\alias{compute_stop_from_start}
\title{Compute the stop given start}
\usage{
compute_stop_from_start(flattened_pd, target_index)
}
\description{
Computes the index of the last token for which \code{col} needs to be updated.
If the line where the last token sits starts with
a closing parenthesis, the stopping token is the last token on the line
before, otherwise it is the last token on the same line as the closing
parenthesis.
}
