% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialized_tests.R
\name{construct_out}
\alias{construct_out}
\title{Construct *-out.R from a *-in.R}
\usage{
construct_out(in_path)
}
\arguments{
\item{in_path}{A character vector that denotes paths to *-in.R files.}
}
\description{
Construct *-out.R from a *-in.R
}
