% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{construct_tree}
\alias{construct_tree}
\title{Construct paths of a tree object given the paths of *-in.R files}
\usage{
construct_tree(in_paths, suffix = "_tree")
}
\arguments{
\item{in_paths}{Character vector of *-in.R files.}

\item{suffix}{Suffix for the tree object.}
}
\description{
Construct paths of a tree object given the paths of *-in.R files
}
\keyword{internal}
