% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical.R
\name{construct_vertical}
\alias{construct_vertical}
\title{Construct an object of class vertical}
\usage{
construct_vertical(x)
}
\arguments{
\item{x}{A character vector or an object of class "vertical".}
}
\description{
Sole puropse of the class vertical is to have a print method that
aligns the output vertically.
}
