% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-is.R
\name{contains_else_expr_that_needs_braces}
\alias{contains_else_expr_that_needs_braces}
\title{Check whether an else expression needs braces}
\usage{
contains_else_expr_that_needs_braces(pd)
}
\arguments{
\item{pd}{A parse table}
}
\description{
Checks whether an else expression in a nest needs braces. Note that for
if-else-if expressions, there is no need to add braces since the if in
else-if will be visited separately with the visitor. This applies to all
conditional statements with more than one alternative.
}
\keyword{internal}
