% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit.R
\name{context_to_terminals}
\alias{context_to_terminals}
\title{Propagate context to terminals}
\usage{
context_to_terminals(pd_nested, passed_lag_newlines, passed_indent,
  passed_spaces)
}
\arguments{
\item{pd_nested}{A nested parse table.}

\item{passed_lag_newlines}{The lag_newlines to be propagated inwards.}

\item{passed_indent}{The indention depth to be propagated inwards.}

\item{passed_spaces}{The number of spaces to be propagated inwards.}
}
\value{
An updated parse table.
}
\description{
Implements a very specific pre-visiting scheme, namely to propagate
indention, spaces and lag_newlines to inner token to terminals. This means
that information regarding indention, linebreaks and spaces (which is
relative in \code{pd_nested}) will be converted into absolute.
}
\seealso{
context_towards_terminals visitors
}
