% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_newlines_to_linebreaks}
\alias{convert_newlines_to_linebreaks}
\title{Replace the newline character with a line break}
\usage{
convert_newlines_to_linebreaks(text)
}
\arguments{
\item{text}{A character vector}
}
\description{
Replace the newline character with a line break
}
\examples{
ensure_newline_is_linebreak("x\\n2")
ensure_newline_is_linebreak(c("x", "2"))
}
\keyword{internal}
