% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{copy_artifacts}
\alias{copy_artifacts}
\title{Copy some file to the test directory that must be present, but are not
passed to the hook as a file argument.}
\usage{
copy_artifacts(artifacts, tempdir)
}
\arguments{
\item{artifacts}{Artifacts to copy.}

\item{tempdir}{The temporary directory.}
}
\description{
Copy some file to the test directory that must be present, but are not
passed to the hook as a file argument.
}
\keyword{internal}
