% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{copy_to_tempdir}
\alias{copy_to_tempdir}
\title{Copy a file to a temporary directory}
\usage{
copy_to_tempdir(path_perm = testthat_file())
}
\arguments{
\item{path_perm}{The path of the file to copy.}
}
\description{
Takes the path to a file as input and returns the path where the temporary
file is stored. Don't forget to unlink once you are done.
}
\keyword{internal}
