% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrupt.R
\name{corrupt_file}
\alias{corrupt_file}
\title{Corrupt a file}
\usage{
corrupt_file(path, extend = 0.3, flat = FALSE,
  transformers = get_transformers(flat = flat))
}
\arguments{
\item{path}{A path to a file to style.}

\item{extend}{The extend of permutation. Zero means no permutation, 1
means permute all tokens in one pool, 0.5 for example means permute
the first half of the token among themselves and the second half of the
token among themselves.}

\item{flat}{Whether to do the styling with a flat approach or with a nested
approach.}

\item{transformers}{A list with functions to be applied to the parsed data.}
}
\description{
Randomize token position before serialization to make people wonder.
}
