% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsed.R
\name{create_filler}
\alias{create_filler}
\title{Enrich parse table with space and linebreak information}
\usage{
create_filler(data)
}
\arguments{
\item{data}{a parse table.}
}
\value{
a parse table with two new columns: newlines and spaces.
}
\description{
This function computes difference (as column and line difference) between two
entries in the parse table and adds this information to the table.
}
\keyword{internal}
