% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.R
\name{create_filler_nested}
\alias{create_filler_nested}
\title{Enrich nested parse table with space and linebreak information}
\usage{
create_filler_nested(pd)
}
\arguments{
\item{pd}{A nested parse table.}
}
\value{
A nested parse table with two new columns: newlines and spaces.
}
\description{
Uses \code{\link[=create_filler]{create_filler()}} in a recursion add space and line break information
separately on every level of nesting.
}
\seealso{
\code{\link[=create_filler]{create_filler()}}
}
