% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_to_tree.R
\name{create_node_from_nested}
\alias{create_node_from_nested}
\title{Create node from nested parse data}
\usage{
create_node_from_nested(pd_nested, parent)
}
\arguments{
\item{pd_nested}{A nested tibble.}

\item{parent}{The parent of the node to be created.}
}
\description{
Create node from nested parse data
}
