% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_to_tree.R
\name{create_node_from_nested_root}
\alias{create_node_from_nested_root}
\title{Convert a nested tibble into a node tree}
\usage{
create_node_from_nested_root(pd_nested)
}
\arguments{
\item{pd_nested}{A nested tibble.}
}
\value{
An object of class "Node" and "R6".
}
\description{
This function is convenient to display all nesting levels of a nested tibble
at once.
}
\examples{
library("magrittr")
code <- "a <- function(x) { if(x > 1) { 1+1 } else {x} }"
l1 <- styler:::compute_parse_data_nested(code) \%>\%
  styler:::create_filler_nested() \%>\%
  styler:::create_node_from_nested_root()
}
