% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-create.R
\name{create_parent_id}
\alias{create_parent_id}
\title{Create the parent id for a new token}
\usage{
create_parent_id(pd_nested)
}
\arguments{
\item{pd_nested}{A nested parse table.}
}
\description{
Just wraps \code{\link[=find_parent]{find_parent()}}, since it can also be used to obtain the parent id
that is needed to create a new token in this nest.
}
