% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-create.R
\name{create_pos_id}
\alias{create_pos_id}
\title{Create a valid pos_id if possible}
\usage{
create_pos_id(pd, pos, after = FALSE, by = 0.1)
}
\arguments{
\item{pd}{A parse table.}

\item{pos}{The position where the new id should be inserted.}

\item{after}{Boolean indicating whether it should be inserted after or before
\code{pos}.}

\item{by}{By how much the reference \code{pos_id} should be increased / decreased
to create a new id.}
}
\value{
Returns a valid pos_id or an error if it was not possible to create one.
}
\description{
Create a valid pos_id if possible
}
