% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_to_tree.R
\name{create_tree}
\alias{create_tree}
\title{Create a tree from text}
\usage{
create_tree(text)
}
\arguments{
\item{text}{A character vector.}
}
\value{
A data frame.
}
\description{
Create a tree representation from a text.
}
