% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{default_style_guide_attributes}
\alias{default_style_guide_attributes}
\title{Enrich parse table with space and line break information}
\usage{
default_style_guide_attributes(pd_flat)
}
\arguments{
\item{pd_flat}{A parse table.}
}
\description{
This function computes difference (as column and line difference) between two
entries in the parse table and adds this information to the table.
}
\examples{
string_to_format <- "call( 3)"
pd <- styler:::compute_parse_data_nested(string_to_format)
styler:::pre_visit(pd, c(default_style_guide_attributes))
}
