% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{delete_if_cache_directory}
\alias{delete_if_cache_directory}
\title{Delete a cache or temp directory}
\usage{
delete_if_cache_directory(path)
}
\arguments{
\item{path}{Absolute path to a directory to delete.}
}
\value{
\code{TRUE} if anything was deleted, \code{FALSE} otherwise.
}
\description{
For safety, \code{path} is only deleted if it is a sub-directory of a temporary
directory or user cache. Since this function relies on \code{tools::R_user_dir()},
it early returns \code{FALSE} on \verb{R < 4.0.0}.
}
\keyword{internal}
