% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-files.R
\name{derive_segment_id}
\alias{derive_segment_id}
\title{Derive the roxygen segment id}
\usage{
derive_segment_id(text, roxygen_examples)
}
\arguments{
\item{text}{Roxygen comments}

\item{roxygen_examples}{Integer sequence that indicates which lines in \code{text}
are roxygen examples. Most conveniently obtained with
\link{identify_start_to_stop_of_roxygen_examples_from_text}.}
}
\description{
This is necessary because technically, two \code{@examples} segments can follow
after each other, and they need to be treated as independent segments.
}
\keyword{internal}
