% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{diff_requires_run_roxygenize}
\alias{diff_requires_run_roxygenize}
\title{Check if we should run roxygen.}
\usage{
diff_requires_run_roxygenize(path_root = here::here())
}
\arguments{
\item{path_root}{The root of project.}
}
\description{
This is the case if a new or replaced/removed line contains a roxygen2
comment in a file that is staged.
}
