% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-files.R
\name{dir_without_._one}
\alias{dir_without_._one}
\title{\code{dir()}, but with full names, ignored case, and included hidden files and
recursive.}
\usage{
dir_without_._one(path, recursive, ...)
}
\description{
\code{dir()}, but with full names, ignored case, and included hidden files and
recursive.
}
\keyword{internal}
