% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dirs_R.cache}
\alias{dirs_R.cache}
\title{Create the path to the precommit R.cache cache}
\usage{
dirs_R.cache(hook_id)
}
\arguments{
\item{hook_id}{The id of the hook for which we want the relative cache
directory.}
}
\description{
This function is only exported for use in hook scripts, but it's not intended
to be called by the end-user directly.
}
\seealso{
Other hook script helpers: 
\code{\link{diff_requires_run_roxygenize}()},
\code{\link{may_require_permanent_cache}()},
\code{\link{roxygen_assert_additional_dependencies}()},
\code{\link{roxygenize_with_cache}()}
}
\concept{hook script helpers}
