% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsed.R
\name{enhance_parse_data}
\alias{enhance_parse_data}
\title{Pre-processing parse data}
\usage{
enhance_parse_data(parse_data)
}
\arguments{
\item{parse_data}{a parse table.}
}
\value{
a pre-processed parse table.
}
\description{
Modifies the parse table minimally by applying some pre-processing steps.
}
\section{Details}{

Preprocessing includes
\itemize{
\item removing non-terminal entries.
\item removing columns id, parent and terminal.
\item adding a start token.
\item adding line-break and space information.
\item removing spaces in comments at the end of the line.
}
}

