% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{ensure_correct_str_txt}
\alias{ensure_correct_str_txt}
\title{Ensure a correct \code{text} of all strings}
\usage{
ensure_correct_str_txt(pd, text)
}
\arguments{
\item{pd}{A parse table.}

\item{text}{The text from which \code{pd} was created. Needed potentially
for another round of parsing.}
}
\description{
Make sure \code{text} of the tokens \code{STR_CONST} is correct and adapt if necessary.
We first parse \code{text} again and include also non-terminal text. Then, we
replace offending \code{text} in the terminal expressions with the text of their
parents if their line / col position matches and return an error otherwise.
}
\keyword{internal}
