% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{ensure_correct_txt}
\alias{ensure_correct_txt}
\title{Ensure a correct \code{text} of all strings and numeric constants}
\usage{
ensure_correct_txt(pd, text)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
Make sure \code{text} of the tokens \code{STR_CONST} and \code{NUM_CONST} is correct and
adapt if necessary. We replace offending \code{text} in the terminal expressions
with the text of their parents if their line / col position matches and
return an error otherwise.
}
\keyword{internal}
