% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_last_is_empty}
\alias{ensure_last_is_empty}
\title{Ensure there is one (and only one) blank line at the end of a vector}
\usage{
ensure_last_is_empty(x)
}
\description{
Ensure there is one (and only one) blank line at the end of a vector
}
\examples{
styler:::ensure_last_is_empty("")
styler:::ensure_last_is_empty(letters)
styler:::ensure_last_is_empty(c(letters, "", "", ""))
}
\keyword{internal}
