% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_named}
\alias{ensure_named}
\title{Name the input}
\usage{
ensure_named(x, candidate_name = NULL, f = identity)
}
\arguments{
\item{x}{A vector.}

\item{f}{How to transform the input \code{x} into a name.}
}
\description{
Name the input
}
\keyword{internal}
