% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-guides.R
\name{equals_style}
\alias{equals_style}
\title{Equals assignment style}
\usage{
equals_style()
}
\description{
Use equals assignment instead of arrow assignment.
}
\details{
This style guide is the same as \code{\link[=tidyverse_style]{tidyverse_style()}}, except it uses
equals (\code{=}) rather than arrow \code{<-} assignment.
}
\examples{
style_text("x <- 1", style = equals_style)
}
\seealso{
Other style_guides: \code{\link{tidyverse_style}}
}
\concept{style_guides}
