% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-files.R
\name{expressions_are_identical}
\alias{expressions_are_identical}
\title{Check whether two expressions are identical}
\usage{
expressions_are_identical(old_text, new_text)
}
\arguments{
\item{old_text}{The initial expression in its character representation.}

\item{new_text}{The styled expression in its character representation.}
}
\description{
Check whether two expressions are identical
}
\keyword{internal}
