% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{extract_diff_root}
\alias{extract_diff_root}
\title{Extract old and new lines from \verb{git diff --cached}}
\usage{
extract_diff_root(root = here::here())
}
\arguments{
\item{root}{The root of project.}
}
\description{
This is useful to detect within a hook script if the core function from a
hook such as \code{\link[roxygen2:roxygenize]{roxygen2::roxygenize()}} must run at all or not.
}
\keyword{internal}
