% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{extract_neighbourhood}
\alias{extract_neighbourhood}
\title{Extract text around the highlighted text}
\usage{
extract_neighbourhood(text, context)
}
\arguments{
\item{text}{Character vector that contains highlighted area.}

\item{context}{The context from \code{\link[=find_active_context]{find_active_context()}}.}
}
\value{
A character vector of length three.
\itemize{
\item First element corresponds to the text on the start line that was not
selected.
\item Second element to the text on the end line that was not selected.
\item Third element is the lines that are not overlapping with the
highlighted text.
}
}
\description{
Extracts unselected text on highlighted lines and text on other lines
}
