% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit.R
\name{extract_terminals_helper}
\alias{extract_terminals_helper}
\title{Helper to extract terminals}
\usage{
extract_terminals_helper(pd_nested)
}
\arguments{
\item{pd_nested}{A nested parse table.}
}
\description{
Helper to extract terminals
}
