% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{find_active_context}
\alias{find_active_context}
\title{Helper to extract and preprocess relevant attributes from
\code{\link[rstudioapi:getActiveDocumentContext]{rstudioapi::getActiveDocumentContext()}}.}
\usage{
find_active_context()
}
\description{
Helper to extract and preprocess relevant attributes from
\code{\link[rstudioapi:getActiveDocumentContext]{rstudioapi::getActiveDocumentContext()}}.
}
