% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-block.R
\name{find_blank_lines_to_next_block}
\alias{find_blank_lines_to_next_block}
\title{Number of lines between cache blocks}
\usage{
find_blank_lines_to_next_block(pd)
}
\arguments{
\item{pd}{A top level nest.}
}
\description{
This is relevant when putting expressions together into a block and preserve
blank lines between them. Note that because code does not need to start on
line 1, the first element of the output is the number of lines until the
first block.
}
\keyword{internal}
