% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-block.R
\name{find_blank_lines_to_next_expr}
\alias{find_blank_lines_to_next_expr}
\title{Find blank lines}
\usage{
find_blank_lines_to_next_expr(pd_nested)
}
\arguments{
\item{pd_nested}{A nested parse table.}
}
\value{
The line number on which the first token occurs.
}
\description{
What number of line breaks lay between the expressions?
}
\keyword{internal}
