% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-roxygen-examples.R
\name{find_dontrun_seqs}
\alias{find_dontrun_seqs}
\title{Find dontrun sequences}
\usage{
find_dontrun_seqs(bare)
}
\arguments{
\item{bare}{Bare code.}
}
\description{
Returns the indices of the lines that correspond to a \code{dontrun} sequence.
}
\keyword{internal}
