% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-line-breaks.R
\name{find_line_break_position_in_multiline_call}
\alias{find_line_break_position_in_multiline_call}
\title{Find index of the token before which the line should be broken}
\usage{
find_line_break_position_in_multiline_call(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
Given a multi-line function call parse table, this function finds the
position of the first named argument and breaks returns the index of it.
If there is no named argument, the line is broken right after the opening
parenthesis.
}
\keyword{internal}
