% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-create.R
\name{find_parent}
\alias{find_parent}
\title{Find the parent of a nest}
\usage{
find_parent(pd_nested)
}
\arguments{
\item{pd_nested}{A nested parse table.}
}
\description{
It is any id of the nest (nested parse table at one level
of nesting) since by definition, all tokens on a nest have the same id.
}
