% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-replacement.R
\name{find_pos_argument}
\alias{find_pos_argument}
\title{Find the index of a positional argument
Assumes that \code{pd} is a function call, which can be tested with
\code{\link[=is_function_call]{is_function_call()}} for example.}
\usage{
find_pos_argument(pd, arg_name)
}
\arguments{
\item{pd}{A parse table}

\item{arg_name}{The name of an argument.}
}
\value{
Returns the index of the argument in the parse table.
}
\description{
Find the index of a positional argument
Assumes that \code{pd} is a function call, which can be tested with
\code{\link[=is_function_call]{is_function_call()}} for example.
}
\keyword{internal}
