% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{find_start_line}
\alias{find_start_line}
\title{Get the start right}
\usage{
find_start_line(pd_nested)
}
\arguments{
\item{pd_nested}{A nested parse table.}
}
\value{
The line number on which the first token occurs.
}
\description{
On what line does the first token occur?
}
