% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-tokens.R
\name{fix_quotes}
\alias{fix_quotes}
\title{Replace single quotes with double quotes}
\usage{
fix_quotes(pd_flat)
}
\arguments{
\item{pd_flat}{A flat parse table.}
}
\description{
Replace single quotes with double quotes
}
\examples{
style_text("'here
is a string
'")
}
\keyword{internal}
