% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.R
\name{flatten_operators}
\alias{flatten_operators}
\title{Flatten some token in the nested parse table based on operators}
\usage{
flatten_operators(pd_nested)
}
\arguments{
\item{pd_nested}{A nested parse table to partially flatten.}
}
\description{
Certain tokens are not placed optimally in the nested parse data with
\code{\link[=compute_parse_data_nested]{compute_parse_data_nested()}}. For example, the token of arithmetic
operations 1 + 1 + 1 should all be on the same level of nesting since the
indention is the same for all but the first two terminals. Setting the
indention correctly is easier to achieve if they are put on the same level of
nesting.
}
\keyword{internal}
