% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.R
\name{flatten_operators_one}
\alias{flatten_operators_one}
\title{Flaten one level of nesting with its child}
\usage{
flatten_operators_one(pd_nested)
}
\arguments{
\item{pd_nested}{A nested parse table.}
}
\description{
Flattening is done in two ways. We can flatten a parse table by moving
the left hand token of an operator one level up. Or doing that with the
right hand token.
}
