% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{force_regex_indention}
\alias{force_regex_indention}
\title{Force indention of tokens that match regex}
\usage{
force_regex_indention(flattened_pd, regex, indention = 0,
  comments_only = TRUE)
}
\description{
Force the level of indention of tokens whose text matches a regular
expression pattern to be a certain amount of spaces. The rule
is only active for the first tokens on a line.
}
