% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{generate_uninstalled_pkg_name}
\alias{generate_uninstalled_pkg_name}
\title{Generate a random package name that is not installed}
\usage{
generate_uninstalled_pkg_name(n = 10)
}
\arguments{
\item{n}{The number of times we should try}
}
\description{
Generate a random package name that is not installed
}
\keyword{internal}
