% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{get_addins_style_name}
\alias{get_addins_style_name}
\alias{get_addins_style_fun}
\title{Return the style function or name}
\usage{
get_addins_style_name()

get_addins_style_fun()
}
\description{
Return the style function or name
}
\keyword{internal}
