% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-code.R
\name{get_engine_pattern}
\alias{get_engine_pattern}
\title{What's the engine pattern for rmd code chunks?}
\usage{
get_engine_pattern()
}
\description{
The function returns the regular expression pattern that identifies
all r engines in Rmd chunks. Defaults to \verb{[Rr]}. You probably only want to
change this if you create a knitr engine that processes R code but is not
the default engine \code{r}.
The pattern must be followed by a space (in the case the chunk is given
a name), a comma (if no name is given but further options are passed to the
engine) or a closing curly brace (in case no option and no name is given to
the chunk).
}
\keyword{internal}
