% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transformers.R
\name{get_transformers}
\alias{get_transformers}
\title{Get the transformer functions for styling}
\usage{
get_transformers(flat, ...)
}
\arguments{
\item{flat}{Whether the transformer functions for flat or nested styling
should be returned.}

\item{...}{Parameters passed to
\itemize{
\item \code{\link[=get_transformers_flat]{get_transformers_flat()}} if \code{flat = TRUE} or
\item \code{\link[=get_transformers_nested]{get_transformers_nested()}} if \code{flat = FALSE}.
}}
}
\value{
A list of transformer functions that operate on flat parse tables.
}
\description{
Get the transformer functions for styling
}
