% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transformers.R
\name{get_transformers_flat}
\alias{get_transformers_flat}
\title{Get the transformer functions for flat styling}
\usage{
get_transformers_flat(strict = TRUE)
}
\arguments{
\item{strict}{A logical value indicating whether a set of strict
or not so strict transformer functions should be returned.}
}
\value{
A list of transformer functions that operate on flat parse
tables.
}
\description{
Get the transformer functions for flat styling
}
\seealso{
Other obtain transformers: \code{\link{get_transformers_nested}}
}
