% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-alignment.R
\name{has_correct_spacing_around_comma}
\alias{has_correct_spacing_around_comma}
\title{Check if spacing around comma is correcr}
\usage{
has_correct_spacing_around_comma(pd_sub)
}
\arguments{
\item{pd_sub}{The subset of a parse table corresponding to one line.}
}
\description{
At least one space after comma, none before, for all but the last comma on
the line
}
\keyword{internal}
