% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{has_crlf_as_first_line_sep}
\alias{has_crlf_as_first_line_sep}
\title{Check if a string uses CRLF EOLs}
\usage{
has_crlf_as_first_line_sep(message, initial_text)
}
\arguments{
\item{message}{A message returned with \code{tryCatch()}.}

\item{initial_text}{The initial text to style.}
}
\description{
Check if a string uses CRLF EOLs
}
\keyword{internal}
