% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{hash_standardize}
\alias{hash_standardize}
\title{Make sure text after styling results in the same hash as text before styling
if it is indeed identical.}
\usage{
hash_standardize(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Make sure text after styling results in the same hash as text before styling
if it is indeed identical.
}
\keyword{internal}
