% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{hook_state_assert}
\alias{hook_state_assert}
\title{Check if the hook produced what you want}
\usage{
hook_state_assert(
  path_candidate,
  tempdir,
  path_candidate_temp,
  file_transformer,
  path_stdout,
  path_stderr,
  expect_success,
  std_err,
  std_out,
  exit_status
)
}
\description{
Match the resulting state after the hook run with the expected state
}
\keyword{internal}
