% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{hooks}
\alias{hooks}
\title{The Hook configuration}
\format{
An object of class \code{list} of length 10.
}
\usage{
hooks
}
\description{
The file \code{.pre-commit-hooks.yaml} is the internal configuration file
(as opposed to \code{.pre-commit-config.yaml} that the user needs). Most users
won't ever need this.
}
\keyword{datasets}
