% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_rmd.R
\name{identify_chunks}
\alias{identify_chunks}
\title{Identify R and text chunks within an Rmd}
\usage{
identify_chunks(lines)
}
\arguments{
\item{lines}{a character vector of lines from an Rmd file}
}
\description{
Identify R and text chunks within an Rmd
}
