% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-code.R
\name{identify_rnw_raw_chunks}
\alias{identify_rnw_raw_chunks}
\title{Identifies raw Rnw code chunks}
\usage{
identify_rnw_raw_chunks(lines)
}
\arguments{
\item{lines}{a character vector of lines from an Rmd file}
}
\description{
Raw in the sense that these chunks don't contain pure R code, but they
contain the Rnw chunk header and footer. All Rnw code chunks are assumed
to contain R code.
}
\keyword{internal}
