% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-roxygen-examples.R
\name{identify_start_stop_of_roxygen_examples_from_text}
\alias{identify_start_stop_of_roxygen_examples_from_text}
\title{Figure out where code examples start and stop}
\usage{
identify_start_stop_of_roxygen_examples_from_text(text)
}
\description{
Finds the start and stop indices of the lines in \code{text} that are
code examples in roxygen comments.
}
