% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-examples.R
\name{identify_start_to_stop_of_roxygen_examples_from_text}
\alias{identify_start_to_stop_of_roxygen_examples_from_text}
\title{Figure out where code examples start and stop}
\usage{
identify_start_to_stop_of_roxygen_examples_from_text(text)
}
\arguments{
\item{text}{A text consisting of code and/or roxygen comments.}
}
\description{
Finds the sequence from start to stop of the lines in \code{text} that are
code examples in roxygen comments.
}
\keyword{internal}
