% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-other.R
\name{if_for_while_part_requires_braces}
\alias{if_for_while_part_requires_braces}
\title{Check if if, for or while loop expression require a braces.}
\usage{
if_for_while_part_requires_braces(pd, key_token)
}
\arguments{
\item{pd}{A parse table.}

\item{key_token}{The token that comes right before the token that contains
the expression to be wrapped (ignoring comments). For if and while loops,
this is the closing "')'", for a for-loop it's "forcond".}
}
\description{
This is the case if they are multi-line and not yet wrapped into curly
braces.
}
\keyword{internal}
