% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize_attributes}
\alias{initialize_attributes}
\alias{initialize_newlines}
\alias{initialize_spaces}
\alias{initialize_multi_line}
\alias{initialize_indention_ref_pos_id}
\alias{initialize_indent}
\alias{validate_parse_data}
\title{Enrich parse table with space and line break information}
\usage{
initialize_attributes(pd_flat)

initialize_newlines(pd_flat)

initialize_spaces(pd_flat)

initialize_multi_line(pd_flat)

initialize_indention_ref_pos_id(pd_flat)

initialize_indent(pd_flat)

validate_parse_data(pd_flat)
}
\arguments{
\item{pd_flat}{A parse table.}
}
\description{
This function computes difference (as column and line difference) between two
entries in the parse table and adds this information to the table.
}
\section{Functions}{
\itemize{
\item \code{initialize_newlines}: Initializes \code{newlines} and \code{lag_newlines}.

\item \code{initialize_spaces}: Initializes \code{spaces}.

\item \code{initialize_multi_line}: Initializes \code{multi_line}.

\item \code{initialize_indention_ref_pos_id}: Initializes \code{indention_ref_ind}.

\item \code{initialize_indent}: Initializes \code{indent}.

\item \code{validate_parse_data}: validates the parse data.
}}

