% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{install_precommit}
\alias{install_precommit}
\title{Install pre-commit on your system.}
\usage{
install_precommit()
}
\description{
This installs pre-commit in the default conda environment of reticulate. It
will be available to use accross different git repositories.
}
