% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{install_precommit_impl}
\alias{install_precommit_impl}
\title{Install pre-commit on your system with conda}
\usage{
install_precommit_impl()
}
\description{
Install pre-commit on your system with conda
}
\keyword{internal}
