% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{is_cached}
\alias{is_cached}
\title{Check if a file is cached}
\usage{
is_cached(x, path_relative_cache)
}
\arguments{
\item{x}{An object, often the last modification time for hooks like
\code{roxygenize}.}

\item{path_relative_cache}{The path of the cache relative to the cache root.}
}
\description{
Check if a file is cached
}
\keyword{internal}
