% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-is.R
\name{is_code_chunk_header}
\alias{is_code_chunk_header}
\title{Identify spinning code chunk header}
\usage{
is_code_chunk_header(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
See https://yihui.name/knitr/demo/stitch/#spin-comment-out-texts for details.
}
\examples{
style_text(c(
  "# title",
  "some_code <- function() {}",
  "#+ chunk-label, opt1=value1",
  "call(3, 2, c(3:2))"
))
}
\keyword{internal}
