% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-is.R
\name{is_code_chunk_header_or_xaringan_or_code_output}
\alias{is_code_chunk_header_or_xaringan_or_code_output}
\title{Identify spinning code chunk header or xaringan}
\usage{
is_code_chunk_header_or_xaringan_or_code_output(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
Wrongly identifies a comment without a preceding line break as a code chunk
header.
See https://yihui.name/knitr/demo/stitch/#spin-comment-out-texts for details.
}
\examples{
style_text(c(
  "# title",
  "some_code <- function() {}",
  "#+ chunk-label, opt1=value1",
  "call(3, 2, c(3:2))",
  "#> 99"
))
}
\keyword{internal}
