% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-indention.R
\name{is_double_indent_function_declaration}
\alias{is_double_indent_function_declaration}
\title{Is the function declaration double indented?}
\usage{
is_double_indent_function_declaration(pd, indent_by = 2L)
}
\arguments{
\item{pd}{A parse table.}

\item{indent_by}{How many spaces of indention should be inserted after
operators such as '('.}
}
\description{
Assumes you already checked if it's a function with
\code{is_function_declaration}. It is double indented if the first token
after the first line break that is a \code{"SYMBOL_FORMALS"}.
}
\keyword{internal}
