% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{is_insufficiently_parsed_strings}
\alias{is_insufficiently_parsed_strings}
\title{Identify strings that were not fully parsed}
\usage{
is_insufficiently_parsed_strings(pd)
}
\arguments{
\item{pd}{A parse table.}

\item{text}{The initial code to style.}
}
\description{
Identifies strings that were not fully parsed due to their vast length.
}
\details{
The meaning of the variable \code{is_problematic_string} in the source code
changes from "all strings" to "all problematic strings", is partly
misleading and this approach was chosen for performance reasons only.
}
\keyword{internal}
