% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-is.R
\name{is_output_prefix}
\alias{is_output_prefix}
\title{Identify output comments}
\usage{
is_output_prefix(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
Determines if parse table contains \verb{#>} at the beginning of a line, which is
often used to prefix the output value of previous R code.
}
\keyword{internal}
