% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-is.R
\name{is_shebang}
\alias{is_shebang}
\title{Identify comments that are shebangs}
\usage{
is_shebang(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
Shebangs should be preserved and no space should be inserted between
\verb{#} and \code{!}. A comment is a shebang if it is the first top-level token
(identified with \code{pos_id}) and if it starts with \verb{#!}.
}
\examples{
style_text("#!/usr/bin/env Rscript")
}
\keyword{internal}
