% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-indention.R
\name{is_single_indent_function_declaration}
\alias{is_single_indent_function_declaration}
\title{Is the function declaration authored with single indentation?}
\usage{
is_single_indent_function_declaration(pd, indent_by = 2L)
}
\arguments{
\item{pd}{A parse table.}

\item{indent_by}{How many spaces of indention should be inserted after
operators such as '('.}
}
\description{
Assumes you already checked if it's a function declaration with
\code{is_function_declaration}. It returns \code{TRUE} if the declaration is authored
with single indentation (where the first argument starts on a new line and
the closing parenthesis \verb{)} starts on a new line, or where arguments on new
lines are indented by \verb{<= 2 * indent_by} spaces). It returns \code{FALSE} if
authored with hanging indentation (where arguments share the line with
\verb{function(} or \verb{)} and are aligned).
}
\keyword{internal}
