% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-is.R
\name{is_tilde_expr}
\alias{is_tilde_expr}
\alias{is_asymmetric_tilde_expr}
\alias{is_symmetric_tilde_expr}
\title{Check whether a parse table contains a tilde}
\usage{
is_tilde_expr(pd, tilde_pos = c(1, 2))

is_asymmetric_tilde_expr(pd)

is_symmetric_tilde_expr(pd)
}
\arguments{
\item{pd}{A parse table.}

\item{tilde_pos}{Integer vector indicating row-indices that should be
checked for tilde. See 'Details'.}
}
\description{
Check whether a parse table contains a tilde
}
\details{
A tilde is on the top row in the parse table if it is an asymmetric tilde
expression (like \code{~column}), in the second row if it is a symmetric tilde
expression (like \code{a~b}).
}
\keyword{internal}
