% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{last_modified}
\alias{last_modified}
\title{Return info of last modified file}
\usage{
last_modified(path_relative_cache, files = NULL)
}
\arguments{
\item{path_relative_cache}{The path of the cache relative to the cache root.}

\item{files}{Character vector with files to check.}
}
\description{
Return info of last modified file
}
\keyword{internal}
