% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-alignment.R
\name{line1_arg_name_and_eq_sub}
\alias{line1_arg_name_and_eq_sub}
\title{Computes the \code{line1} attribute for the first two tokens of the of a sub
parse table if it contains an \code{EQ_SUB}.}
\usage{
line1_arg_name_and_eq_sub(pd_sub)
}
\description{
Computes the \code{line1} attribute for the first two tokens of the of a sub
parse table if it contains an \code{EQ_SUB}.
}
