% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{line_break_after}
\alias{line_break_after}
\title{Check whether a line break follows}
\usage{
line_break_after(flattened_pd, target_index)
}
\description{
Check wheter a code line break follows, that is, whether a line break
separates \code{target_index} and the next non-comment element.
}
