% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.R
\name{lookup_new_special}
\alias{lookup_new_special}
\title{lookup which new tokens were created from "SPECIAL"}
\usage{
lookup_new_special(regex = NA)
}
\description{
lookup which new tokens were created from "SPECIAL"
}
