% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{make_transformer}
\alias{make_transformer}
\title{Closure to return a transformer function}
\usage{
make_transformer(transformers)
}
\arguments{
\item{transformers}{A list of transformer functions.
TODO:
\itemize{
\item Walk tree defined by \code{leaves}, compute whitespace
\item Perform all transformations on hierarchical structure
\itemize{
\item Compute text for a sub-element
}
\item Compute indentation
\itemize{
\item Braces
\item Function calls
\item Function definitions
}
\item Remove \code{includeText = TRUE}
}}
}
\description{
This function takes a list of transformer functions as input and
returns a function that can be applied to character strings
that should be transformed.
}
