% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{make_transformer_flat}
\alias{make_transformer_flat}
\title{A Closure to return transformer function}
\usage{
make_transformer_flat(transformers = transformers)
}
\arguments{
\item{transformers}{A list of transformer functions that operate on flat
parse tables.}
}
\description{
Returns a closure that turns \code{text} into a flat parse table and applies
\code{transformers}  on it.
}
\seealso{
Other make transformers: \code{\link{make_transformer_nested}},
  \code{\link{make_transformer}}
}
