% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-files.R
\name{map_filetype_to_pattern}
\alias{map_filetype_to_pattern}
\title{Map the file type to a corresponding regular expression}
\usage{
map_filetype_to_pattern(filetype)
}
\arguments{
\item{filetype}{The file type to map to a regex.}
}
\description{
Map the file type to a corresponding regular expression
}
\examples{
styler:::map_filetype_to_pattern(c(".rMd", "R"))
}
\keyword{internal}
