% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-examples.R
\name{match_stop_to_start}
\alias{match_stop_to_start}
\title{Match a stop candidate to a start}
\usage{
match_stop_to_start(start, stop_candidates)
}
\arguments{
\item{start}{An integer.}

\item{stop_candidates}{Potential stop candidates.}
}
\description{
Match a stop candidate to a start
}
\examples{
styler:::match_stop_to_start(1, c(3, 4, 5))
}
\keyword{internal}
