% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{may_require_permanent_cache}
\alias{may_require_permanent_cache}
\title{Issue a warning if \code{{R.cache}} uses temporary cache only}
\usage{
may_require_permanent_cache(temporary_cache = FALSE)
}
\arguments{
\item{temporary_cache}{Whether a temporary cache is accepted or not. \code{TRUE}
means no warning will be issued, \code{FALSE} means a warning will be issued if
no permanent cache is available.}
}
\description{
This function is only exported to have a clean API, but it's not intended to
be used by the end-user, only in hook scripts.
}
