% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{n_times_faster_with_cache}
\alias{n_times_faster_with_cache}
\title{Times two function calls with temporarily enabled cache}
\usage{
n_times_faster_with_cache(
  x1,
  x2 = x1,
  ...,
  fun = styler::style_text,
  n = 3,
  clear = "always"
)
}
\arguments{
\item{...}{Arguments passed to \code{fun}.}

\item{fun}{The function that should be timed.}

\item{n}{The number of times the experiment should be repeated.}
}
\value{
A scalar indicating the relative difference of the second compared to the
first run.
}
\description{
This can be helpful for benchmarking.
}
\keyword{internal}
