% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent.R
\name{needs_indention_one}
\alias{needs_indention_one}
\title{Check whether indention is needed}
\usage{
needs_indention_one(pd, potential_trigger_pos, other_trigger_tokens)
}
\arguments{
\item{pd}{A parse table.}

\item{potential_trigger_pos}{the index of the token in the parse table
for which it should be checked whether it should trigger indention.}

\item{other_trigger_tokens}{Other tokens that are going to cause indention
if on the same line as the token corresponding to \code{potential_trigger}.}
}
\value{
Returns \code{TRUE} if indention is needed, \code{FALSE} otherwise.

\code{TRUE} if indention is needed, \code{FALSE} otherwise.
}
\description{
Indention is needed if the two conditions apply:
}
\details{
\itemize{
\item there is no multi-line token between the trigger and the first line break.
\item there is no other token between the potential trigger and the first line
break that is going to cause indention.
}
}
\keyword{internal}
