% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-examples-parse.R
\name{needs_rd_emulation}
\alias{needs_rd_emulation}
\title{Check if rd emulation is required with \code{\link[roxygen2:roc_proc_text]{roxygen2::roc_proc_text()}}}
\usage{
needs_rd_emulation(roxygen)
}
\description{
Check if rd emulation is required with \code{\link[roxygen2:roc_proc_text]{roxygen2::roc_proc_text()}}
}
\keyword{internal}
