% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{needs_reindention}
\alias{needs_reindention}
\title{Check whether an expression needs reindention}
\usage{
needs_reindention(flattened_pd, target_index)
}
\arguments{
\item{flattened_pd}{A flattened parse table.}

\item{target_index}{The index of the token in \code{flattened_pd} that has the
indention we want to apply to other tokens.}

\item{cols_to_update}{The indices of the tokens in the parse table which
should be shifted.}
}
\description{
Check whether an opening round brace is on the same line. If not, check
whether a curly opening brace is there or whether there is nothing following
on the same line at all. If so, there is no need to indent.
}
\details{
If there is an opening round brace on the same line, check whether the
closing brace is on that very same line too. If so, re-indention is needed,
otherwise not.
}
