% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{newlines_and_spaces}
\alias{newlines_and_spaces}
\title{concentrate newlines an spaces in a string}
\usage{
newlines_and_spaces(newlines, spaces)
}
\arguments{
\item{newlines}{Scalar indicating how many newlines ("\ n") should returned.}

\item{spaces}{Scalar indicating how many spaces should be appended to the
newlines.}
}
\value{
A string.
}
\description{
concentrate newlines an spaces in a string
}
