% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-navigate-nest.R
\name{next_non_comment}
\alias{next_non_comment}
\alias{previous_non_comment}
\title{Find the index of the next or previous non-comment in a parse table.}
\usage{
next_non_comment(pd, pos)

previous_non_comment(pd, pos)
}
\arguments{
\item{pd}{A parse table.}

\item{pos}{The position of the token to start the search from.}
}
\description{
Find the index of the next or previous non-comment in a parse table.
}
\seealso{
Other third-party style guide helpers: 
\code{\link{pd_is}},
\code{\link{scope_normalize}()}
}
\concept{third-party style guide helpers}
\keyword{internal}
