% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{next_terminal}
\alias{next_terminal}
\title{If the first is a terminal, return it. If not, go inside it and search the
next terminal}
\usage{
next_terminal(pd, stack = FALSE, vars = c("pos_id", "token", "text"),
  exclude_tokens = list())
}
\arguments{
\item{pd}{A nest.}

\item{stack}{Whether or not to stack non-terminals. The order is inside-out,
i.e. the first non-terminal on top, the terminal last.}
}
\description{
If the first is a terminal, return it. If not, go inside it and search the
next terminal
}
\examples{
pd <- compute_parse_data_nested("if (TRUE) f()")
styler::next_terminal(pd)
}
\keyword{internal}
