% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{not_conda}
\alias{not_conda}
\title{The testing environment does not use a conda environment if the env variable
PRECOMMIT_INSTALLATION_METHOD is not 'conda'.}
\usage{
not_conda()
}
\description{
The testing environment does not use a conda environment if the env variable
PRECOMMIT_INSTALLATION_METHOD is not 'conda'.
}
\keyword{internal}
