% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{option_read}
\alias{option_read}
\title{Get the value of an option}
\usage{
option_read(x, default = NULL, error_if_not_found = TRUE)
}
\arguments{
\item{x}{a character string holding an option name.}

\item{default}{if the specified option is not set in the options list,
    this value is returned.  This facilitates retrieving an option and
    checking whether it is set and setting it separately if not.}

\item{error_if_not_found}{Whether or not an error should be returned if the
option was not set.}
}
\description{
Basically a \code{getOptions()} that fails fast by default.
}
\keyword{internal}
