% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-examples-parse.R
\name{parse_roxygen}
\alias{parse_roxygen}
\title{Parse roxygen comments into text}
\usage{
parse_roxygen(roxygen)
}
\arguments{
\item{roxygen}{Roxygen comments.}
}
\description{
Used to parse roxygen code examples
}
\examples{
styler:::parse_roxygen(
"#' @examples
 #' 1+  1
")
}
\keyword{internal}
