% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-files.R
\name{parse_transform_serialize}
\alias{parse_transform_serialize}
\title{Parse, transform and serialize text}
\usage{
parse_transform_serialize(text, transformers)
}
\arguments{
\item{text}{A character vector to parse.}

\item{transformers}{A list of \emph{named} transformer functions}
}
\description{
Wrapper function for the common three operations.
}
\keyword{internal}
