% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-files.R
\name{parse_transform_serialize_r}
\alias{parse_transform_serialize_r}
\title{Parse, transform and serialize text}
\usage{
parse_transform_serialize_r(
  text,
  transformers,
  base_indention,
  warn_empty = TRUE
)
}
\arguments{
\item{text}{The text to parse.}

\item{transformers}{Passed to \code{\link[=cache_make_key]{cache_make_key()}} to generate a key.}

\item{base_indention}{Integer scalar indicating by how many spaces the whole
output text should be indented. Note that this is not the same as splitting
by line and add a \code{base_indention} spaces before the code in the case
multi-line strings are present. See 'Examples'.}

\item{warn_empty}{Whether or not a warning should be displayed when \code{text}
does not contain any tokens.}
}
\description{
Wrapper function for the common three operations.
}
\seealso{
\code{\link[=parse_transform_serialize_roxygen]{parse_transform_serialize_roxygen()}}
}
\keyword{internal}
