% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-files.R
\name{parse_transform_serialize_roxygen}
\alias{parse_transform_serialize_roxygen}
\title{Parse, transform and serialize roxygen comments}
\usage{
parse_transform_serialize_roxygen(text, transformers)
}
\description{
Splits \code{text} into roxygen code examples and non-roxygen code examples and
then maps over these examples by applyingj
\code{\link[=style_roxygen_code_examples_one_example]{style_roxygen_code_examples_one_example()}}.
}
\keyword{internal}
