% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{path_derive_precommit_exec}
\alias{path_derive_precommit_exec}
\title{Derive the path to the pre-commit executable}
\usage{
path_derive_precommit_exec()
}
\description{
First check if there is an executable on the \verb{$PATH}, then check if the
conda installation was used. Returns \code{""} if no executable is found. If not,
search other possible locations that are OS dependent.
}
\keyword{internal}
