% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{path_derive_precommit_exec_impl}
\alias{path_derive_precommit_exec_impl}
\title{Find an executable}
\usage{
path_derive_precommit_exec_impl(candidate)
}
\arguments{
\item{candidate}{A directory to check for the pre-commit executable. The
directory may also not exist.}
}
\description{
Evaluates if the pre-commit executable exists in one or more candidate
locations. If so, return one, else return the empty string
}
\keyword{internal}
