% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{path_derive_precommit_exec_path}
\alias{path_derive_precommit_exec_path}
\title{Derive the pre-commit executable from the path}
\usage{
path_derive_precommit_exec_path()
}
\description{
Tries to derive the \code{pre-commit} executable from the \verb{$PATH}.
Returns \code{""} if no executable is found.
}
\keyword{internal}
