% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{path_derive_precommit_exec_win_python3plus_base}
\alias{path_derive_precommit_exec_win_python3plus_base}
\title{Where are executables on Windows for Python 3 and higher?}
\usage{
path_derive_precommit_exec_win_python3plus_base()
}
\description{
Heuristic to determine the directory where the pre-commit executable on
Windows lives for Python versions 3 and above.
}
\keyword{internal}
