% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{path_precommit_exec}
\alias{path_precommit_exec}
\alias{path_pre_commit_exec}
\title{Locate the pre-commit executable}
\usage{
path_precommit_exec(check_if_exists = TRUE)

path_pre_commit_exec(check_if_exists = TRUE)
}
\arguments{
\item{check_if_exists}{Whether or not to make sure the returned path also
exists.}
}
\description{
Locate the pre-commit executable
}
