% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent.R
\name{pd_is_multi_line}
\alias{pd_is_multi_line}
\title{Check whether a parse table is a multi-line token}
\usage{
pd_is_multi_line(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
A token is a multi-line expression if and only if:
}
\details{
\itemize{
\item it contains a line break.
\item it has at least one child that is a multi-line expression itself.
}
}
\keyword{internal}
