% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrupt.R
\name{permute}
\alias{permute}
\title{Permute elements in a vector}
\usage{
permute(vec, extend)
}
\arguments{
\item{vec}{A vector to permute.}

\item{extend}{The extend of permutation. Zero means no permutation, 1
means permute all tokens in one pool, 0.5 for example means permute
the first half of the token among themselves and the second half of the
token among themselves.}
}
\description{
Permute elements in a vector
}
