% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-roxygen-examples.R
\name{post_parse_roxygen}
\alias{post_parse_roxygen}
\title{Fix parsing bugs}
\usage{
post_parse_roxygen(raw)
}
\arguments{
\item{raw}{Raw code to post-process.}
}
\description{
Fix parsing bugs
}
\examples{
code <- "style_text('call( 1)')
style_text('1    + 1', strict = FALSE)
style_text('a\%>\%b', scope = 'spaces')
style_text('a\%>\%b; a', scope = 'line_breaks')
style_text('a\%>\%b; a', scope = 'tokens')"
parsed <- styler:::parse_roxygen(code) # cuts before "\%" for no reason
fixed <- styler:::post_parse_roxygen(styler:::drop_newline_codelines(parsed)) # better
}
