% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precommit-package.R
\docType{package}
\name{precommit-package}
\alias{precommit}
\alias{precommit-package}
\title{precommit: Pre commit hooks}
\description{
Useful pre-commit hooks for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lorenzwalthert.github.io/precommit/}
  \item \url{https://github.com/lorenzwalthert/precommit}
}

}
\keyword{internal}
