% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precommithooks-package.R
\docType{package}
\name{precommithooks-package}
\alias{precommithooks}
\alias{precommithooks-package}
\title{precommithooks: Pre commit hooks}
\description{
Useful pre-commit hooks for R.
}
\keyword{internal}
