% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{prettify_ws}
\alias{prettify_ws}
\alias{prettify_parsed}
\title{Prettify R source code}
\usage{
prettify_ws(pkg = ".")

prettify_parsed(pkg = ".")
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package}
}
\description{
Performs various substitutions in all \code{.R} files in a package.
Carefully examine the results after running this function!
}
