% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical.R
\name{print.vertical}
\alias{print.vertical}
\title{Print styled code}
\usage{
\method{print}{vertical}(x, ..., colored = TRUE,
  style = prettycode::default_style())
}
\arguments{
\item{x}{A character vector, one element corresponds to one line of code.}

\item{...}{Not currently used.}

\item{colored}{Whether or not the output should be colored with
\code{prettycode::highlight()}.}

\item{style}{Passed to \code{prettycode::highlight()}.}
}
\description{
Print styled code
}
