% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{rds_to_version}
\alias{rds_to_version}
\title{Convert a serialized R object to a certain version.}
\usage{
rds_to_version(path, version = 2)
}
\arguments{
\item{path}{A path to an rds file.}

\item{version}{The target version.}
}
\description{
Needed to make \code{\link[testthat:expect_known_value]{testthat::expect_known_value()}} work on R < 3.6.
}
\keyword{internal}
