% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.R
\name{re_nest}
\alias{re_nest}
\title{Reorganize a nested parse table}
\usage{
re_nest(pd_nested)
}
\arguments{
\item{pd_nested}{A nested parse table to reorganize.}
}
\description{
Certain tokens are not placed optimally in the nested parse data with
\code{\link[=compute_parse_data_nested]{compute_parse_data_nested()}}. For example, the token of arithmetic
operations 1 + 1 + 1 should all be on the same level of nesting since
the indention is the same for all but the first two terminals. Setting the
indention correcly is easier to achieve if they are put on the same level
of nesting.
}
\examples{
code <- "1+1+1+1"
# tree-structure with compute_parse_data_nested
styler:::create_tree(code, re_nest = FALSE)
# tree-structure with re_nest post-processing
styler:::create_tree(code, re_nest = TRUE)
}
