% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_utf8}
\alias{read_utf8}
\title{Read UTF-8}
\usage{
read_utf8(path)
}
\arguments{
\item{path}{A path to a file to read.}
}
\description{
Reads an UTF-8 file, returning the content and whether or not the final line
was blank. This information is required higher up in the callstack because
we should write back if contents changed or if there is no blank line at the
EOF. A perfectly styled file with no EOF blank line will gain such a line
with this implementation.
}
\keyword{internal}
