% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-line-breaks.R
\name{reduce_extra_blank_lines_between_scopes}
\alias{reduce_extra_blank_lines_between_scopes}
\title{Reduce multiple blank lines to a maximum number of allowed blank lines}
\usage{
reduce_extra_blank_lines_between_scopes(pd_flat, allowed_blank_lines = 2L)
}
\arguments{
\item{pd_flat}{A flat parse table.}

\item{allowed_blank_lines}{The maximum number of allowed blank lines between code elements. Default is \code{2L}.}
}
\description{
Reduce multiple blank lines to a maximum number of allowed blank lines
}
\keyword{internal}
