% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{release_gh}
\alias{release_gh}
\title{Create a new release on GitHub}
\usage{
release_gh(bump = "patch")
}
\description{
This must be done \strong{before} a CRAN release.
}
\details{
\itemize{
\item bump description.
\item update default config in inst/
\item commit
\item git tag
\item run conistent-release-tag hook with --release-mode (passing args to hooks
not possible interactivley, hence we run in advance).
\item commit and push with skipping conistent-release-tag
\item autoupdate own config file
\item bump description with dev
\item commit and push DESCRIPTION and .pre-commit-config.yaml
}
}
