% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.R
\name{relocate_eq_assign}
\alias{relocate_eq_assign}
\title{Relocate the expressions containing the token \code{EQ_ASSIGN} within the nested
parse table}
\usage{
relocate_eq_assign(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
This used to be relevant when \code{{styler}} supported R versions \verb{< 3.5}, and,
therefore, parser version \code{1}, which had a bug that did not produce the same
hierarchy of the parse table (parent and id relationship) for \verb{<-} and \code{=}
(See 'Examples').
}
\examples{
styler:::get_parse_data("a <- b <- 3")
styler:::get_parse_data("a  = b = 3")
styler:::get_parse_data(
  "x = 5
  if(x >= 5)
  y = TRUE else
  y = FALSE",
)
styler:::get_parse_data(
  "x <- 5
  if(x >= 5)
  y <- TRUE else
  y <- FALSE",
)
}
\keyword{internal}
