% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.R
\name{relocate_eq_assign_one}
\alias{relocate_eq_assign_one}
\title{Relocate an assignment expression}
\usage{
relocate_eq_assign_one(pd)
}
\arguments{
\item{pd}{A parse table with one assignment expression to relocate.}
}
\description{
Relocates an assignment expression within a parse table containing one
assignment expression. Note that one assignment can include multiple
assignment operators such as "a = b = c".
}
