% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{remind_removing_old_cache}
\alias{remind_removing_old_cache}
\title{Ask people to remove the cache}
\usage{
remind_removing_old_cache()
}
\description{
The way RStudio Startup works does not allow o read the prompt for some
reasons (https://stackoverflow.com/questions/55772436/readline-does-not-prompt-user-input-from-rprofile-site-in-rstudio)
So we better don't use the prompt and issue a message only.
}
\keyword{internal}
