% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{remove_cache_old_versions}
\alias{remove_cache_old_versions}
\title{Ask people to remove the cache}
\usage{
remove_cache_old_versions()
}
\description{
The way RStudio Startup works does not allow o read the prompt for some
reasons (https://stackoverflow.com/questions/55772436/readline-does-not-prompt-user-input-from-rprofile-site-in-rstudio)
So we better don't use the prompt and issue a message only.
}
\keyword{internal}
