% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-roxygen-examples.R
\name{remove_dont_mask}
\alias{remove_dont_mask}
\title{Remove dont* mask}
\usage{
remove_dont_mask(roxygen)
}
\arguments{
\item{roxygen}{Roxygen code examples that contains a dont* segment only.}
}
\description{
Remove dont* mask
}
\keyword{internal}
