% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-roxygen-examples.R
\name{remove_dontrun_mask}
\alias{remove_dontrun_mask}
\title{Remove dontrun mask}
\usage{
remove_dontrun_mask(roxygen)
}
\arguments{
\item{roxygen}{Roxygen code examples that contains a dontrun segment only.}
}
\description{
Remove dontrun mask
}
\keyword{internal}
