% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-examples-add-remove.R
\name{remove_roxygen_header}
\alias{remove_roxygen_header}
\title{Remove roxygen header}
\usage{
remove_roxygen_header(text)
}
\description{
Can't simply remove the element with the regex because it may happen that
the roxygen tag is on the same line as its contents start.
}
\examples{
#' @examples c(1, 2)
}
\keyword{internal}
