% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-strings.R
\name{rep_char}
\alias{rep_char}
\title{Repeat elements of a character vector \code{times} times and collapse it}
\usage{
rep_char(char, times)
}
\arguments{
\item{char}{A character vector.}

\item{times}{an integer giving the number of repetitions.}
}
\value{
A character vector.
}
\description{
Repeat elements of a character vector \code{times} times and collapse it
}
\keyword{internal}
