% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rev_read}
\alias{rev_read}
\title{Read the refs corresponding to a hooks repo}
\usage{
rev_read(path = ".pre-commit-config.yaml", repo = hooks_repo)
}
\description{
Read the refs corresponding to a hooks repo
}
\keyword{internal}
