% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{roxygen_assert_additinal_dependencies}
\alias{roxygen_assert_additinal_dependencies}
\title{Assert if all dependencies are installed}
\usage{
roxygen_assert_additinal_dependencies()
}
\description{
This function is only exported for use in hook scripts, but it's not intended
to be called by the end-user directly.
}
\seealso{
Other hook script helpers: 
\code{\link{diff_requires_run_roxygenize}()},
\code{\link{may_require_permanent_cache}()}
}
\concept{hook script helpers}
