% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{roxygenize_with_cache}
\alias{roxygenize_with_cache}
\title{Roxygen depending on cache state}
\usage{
roxygenize_with_cache(key, dirs)
}
\description{
This function is only exported for use in hook scripts, but it's not intended
to be called by the end-user directly.
}
\seealso{
Other hook script helpers: 
\code{\link{diff_requires_run_roxygenize}()},
\code{\link{dirs_R.cache}()},
\code{\link{may_require_permanent_cache}()},
\code{\link{roxygen_assert_additinal_dependencies}()}
}
\concept{hook script helpers}
