% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{save_after_styling_is_active}
\alias{save_after_styling_is_active}
\title{Heuristic to see if a file styled with the addin should be saved or not.}
\usage{
save_after_styling_is_active()
}
\description{
Using the R option \code{"styler.save_after_styling"} and if unset, checks legacy
method via environment variable \code{save_after_styling}.
}
\keyword{internal}
