% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-guides.R
\name{scope_normalize}
\alias{scope_normalize}
\title{Convert the styling scope to its lower-level representation}
\usage{
scope_normalize(scope, name = substitute(scope))
}
\arguments{
\item{scope}{A character vector of length one or a vector of class \code{AsIs}.}

\item{name}{The name of the character vector to be displayed if the
construction of the factor fails.}
}
\description{
If \code{scope} is of class \code{character} and of length one, the value of the
argument and all less-invasive levels are included too (e.g.
styling tokens includes styling spaces). If
\code{scope} is of class \code{AsIs}, every level to be included has to be declared
individually. See compare \code{\link[=tidyverse_style]{tidyverse_style()}} for the possible levels and
their order.
}
\examples{
scope_normalize(I("tokens"))
scope_normalize(I(c("indention", "tokens")))
}
\seealso{
Other third-party style guide helpers: 
\code{\link{next_non_comment}()},
\code{\link{pd_is}}
}
\concept{third-party style guide helpers}
