% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-code.R
\name{separate_chunks}
\alias{separate_chunks}
\title{Separate chunks within Rmd and Rnw contents}
\usage{
separate_chunks(lines, filetype = "Rmd")
}
\arguments{
\item{lines}{a character vector of lines from an Rmd file}
}
\description{
Identifies and separates the code and text chunks (the latter includes non-R
code) within an Rmd or Rnw file, and returns these separately.
}
\keyword{internal}
