% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-alignment.R
\name{serialize_column}
\alias{serialize_column}
\title{Serialize all lines for a given column}
\usage{
serialize_column(relevant_pd_by_line, column)
}
\arguments{
\item{relevant_pd_by_line}{A list with parse tables of a multi-line call,
excluding first and last column.}

\item{colum}{Which column to serialize.}
}
\description{
Serialize all lines for a given column
}
\keyword{internal}
