% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-alignment.R
\name{serialize_line}
\alias{serialize_line}
\title{Serialize one line for a column}
\usage{
serialize_line(relevant_pd_by_line, is_last_line, column)
}
\arguments{
\item{relevant_pd_by_line}{A list with parse tables of a multi-line call,
excluding first and last column.}

\item{is_last_line}{Boolean for every element of \code{relevant_pd_by_line}
indicating if it is the last line.}

\item{column}{The index of the column to serialize.}
}
\description{
Serialize one line for a column
}
