% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsed.R, R/serialize.R
\name{serialize_parse_data_flat}
\alias{serialize_parse_data_flat}
\alias{serialize_parse_data_flat}
\title{Serialize Flat Parse Data}
\usage{
serialize_parse_data_flat(pd_flat)

serialize_parse_data_flat(pd_flat)
}
\arguments{
\item{pd_flat}{A parse table.}

\item{pd_flat}{A parse table.}
}
\description{
Collapses a parse table into character vector representation.

Collapses a parse table into character vector representation.
}
\details{
The function essentially collapses the column text of \code{pd_flat}
while taking into account space and linebreak information from the columns
newlines and spaces. \cr
Roughly speaking, this is the inverse operation of
\code{\link[=compute_parse_data_flat_enhanced]{compute_parse_data_flat_enhanced()}}, which turns a character vector into a
parse table, since \code{serialize_parse_data_flat()} turns a parse table back
into a character vector.

The function essentially collapses the column text of \code{pd_flat}
while taking into account space and linebreak information from the columns
newlines and spaces. \cr
Roughly speaking, this is the inverse operation of
\code{\link[=compute_parse_data_flat_enhanced]{compute_parse_data_flat_enhanced()}}, which turns a character vector into a
parse table, since \code{serialize_parse_data_flat()} turns a parse table back
into a character vector.
}
