% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.R
\name{serialize_parse_data_nested}
\alias{serialize_parse_data_nested}
\title{Serialize a nested parse table}
\usage{
serialize_parse_data_nested(pd_nested)
}
\arguments{
\item{pd_nested}{A nested parse table with line break, spaces and indention
information.}
}
\value{
A character string.
}
\description{
Collapses a nested parse table into its character vector representation.
}
