% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-assert-args.R
\name{set_and_assert_arg_filetype}
\alias{set_and_assert_arg_filetype}
\title{Set the file type argument}
\usage{
set_and_assert_arg_filetype(filetype)
}
\arguments{
\item{filetype}{A character vector with file types to convert to the internal
standard format.}
}
\description{
Sets and asserts the file type argument to a standard format for further internal
processing.
}
\examples{
styler:::set_and_assert_arg_filetype("rMd")
try(styler:::set_and_assert_arg_filetype("xyz"))
}
\keyword{internal}
