% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-assert-args.R
\name{set_arg_paths}
\alias{set_arg_paths}
\title{Standardize paths in root}
\usage{
set_arg_paths(path)
}
\arguments{
\item{path}{A path.}
}
\description{
Standardization required to use \code{setdiff()} with paths.
}
\examples{
styler:::set_arg_paths(c("./file.R", "file.R", "../another-file.R"))
}
\seealso{
dir_without_.
}
\keyword{internal}
