% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-args.R
\name{set_args}
\alias{set_args}
\alias{set_arg_write_tree}
\title{Set arguments}
\usage{
set_arg_write_tree(write_tree)
}
\arguments{
\item{write_tree}{Whether or not to write tree.}
}
\description{
Set arguments
}
\section{Functions}{
\itemize{
\item \code{set_arg_write_tree}: Sets the argument \code{write_tree} in
\code{\link[=test_collection]{test_collection()}} to be \code{TRUE} for R versions higher or equal to 3.2, and
\code{FALSE} otherwise since the second-level dependency \code{DiagrammeR} from
\code{data.table} is not available for R < 3.2.
}}

