% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-line-break.R, R/rules-spacing.R
\name{set_line_break_around_curly_curly}
\alias{set_line_break_around_curly_curly}
\alias{set_space_in_curly_curly}
\title{Styling around \code{\{\{}}
\usage{
set_line_break_around_curly_curly(pd)

set_space_in_curly_curly(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
Styling around \code{\{\{}
}
\details{
curly-curly affects styling of line break and spaces, namely:
\itemize{
\item No line break after first or second \code{\{}, before third and fourth \code{\{}.
\item No space after first and third \code{\{}, one space after second and before third
\code{\}}.
}
}
\seealso{
style_text_without_curly_curly
}
\keyword{internal}
