% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_path_cp_config_from}
\alias{set_path_cp_config_from}
\title{Set the location to a config file}
\usage{
set_path_cp_config_from(path_cp_config_from, path_root, verbose = TRUE)
}
\description{
If a remote location is specified, the file is downloaded to a temporary
location and the path to this location is returned. If \code{NULL}, we'll resort
to a default config. We'll perform some checks on the existence of the file
too.
}
\keyword{internal}
