% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-spacing.R
\name{set_space_around_op}
\alias{set_space_around_op}
\title{Set spaces around operators}
\usage{
set_space_around_op(pd_flat, strict)
}
\description{
Alignement is kept, if detected.
}
\keyword{internal}
