% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-spaces.R
\name{set_space_around_operator}
\alias{set_space_around_operator}
\title{Set spaces around operators}
\usage{
set_space_around_operator(pd_flat, strict)
}
\description{
Alignment is kept, if detected.
}
\keyword{internal}
