% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-spaces.R
\name{set_space_between_levels}
\alias{set_space_between_levels}
\title{Set space between levels of nesting}
\usage{
set_space_between_levels(pd_flat)
}
\arguments{
\item{pd_flat}{A flat parse table.}
}
\description{
With the nested approach, certain rules do not have an effect anymore because
of the nature of the nested structure. Setting spacing before curly
brackets in for / if / while statements and function declarations will be
such a case since a curly bracket is always at the first position in a parse
table, so spacing cannot be set after the previous token.
}
\keyword{internal}
