% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{set_token_dependent_indention}
\alias{set_token_dependent_indention}
\title{Make indention token-dependent}
\usage{
set_token_dependent_indention(flattened_pd)
}
\arguments{
\item{flattened_pd}{A flattened parse table.}
}
\description{
Updates the indention of certain tokens to be dependent on the position
of other tokens.
}
\details{
For details of the implmementation, see
\code{\link[=set_token_dependent_indention_one]{set_token_dependent_indention_one()}}.
}
\note{
See 'Examples' for two examples of how formatted code should look like. Note
that the second example is not compliant with the tidyverse style guide.
In this case, the approach to take would be to have a transformer function
/ rule that breaks the line after the opening parenthesis anyways, so no
token-dependent indention would be needed. Nevertheless, the case should be
covered because other style guides may depend on this reformatting and do
not implement a rule that adds a line break after the opening parenthesis.
}
\examples{
\dontrun{
a <- function(x,
              y) {
  x
}

call(1,
     2)
}
}
