% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unindent.R
\name{set_unindention_child}
\alias{set_unindention_child}
\title{Unindent a child if necessary}
\usage{
set_unindention_child(pd, token = "')'", unindent_by)
}
\arguments{
\item{pd}{A parse table.}

\item{token}{The token the unindention should be based on.}

\item{unindent_by}{By how many spaces one level of indention is reversed.}
}
\description{
check whether any of the children of \code{pd} has \code{token} on the same line as the
closing \code{token} of pd. If so, unindent that token.
}
\keyword{internal}
