% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{snippet_generate}
\alias{snippet_generate}
\title{Generate code snippets}
\usage{
snippet_generate(snippet = "", open = interactive(), root = here::here())
}
\arguments{
\item{snippet}{Name of the snippet.}

\item{root}{The path to the root directory of your project.}
}
\description{
Utility function to generate code snippets:
}
\details{
Currently supported:
\itemize{
\item additional-deps-roxygenize: Code to paste into
\code{.pre-commit-config.yaml} for the additional dependencies required by
roxygen2.
}
}
