% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_guides.R
\name{specify_reindention}
\alias{specify_reindention}
\title{Specify what is re-indented how}
\usage{
specify_reindention(regex_pattern = regex_strcode(), indention = 0,
  comments_only = TRUE)
}
\arguments{
\item{regex_pattern}{Character vector with regular expression patterns that
are to be re-indented with spaces.}

\item{indention}{The indention tokens should have if they match
\code{regex_pattern}.}

\item{comments_only}{Whether the \code{regex_reindention_pattern} should only be
matched against comments or against all tokens. Mainly added for
performance.}
}
\description{
This function returns a list that can be used as an input for the argument
\code{reindention} of the function \code{\link[=tidyverse_style]{tidyverse_style()}}. It features sensible
defaults, so the user can specify deviations from them conveniently without
the need of setting all arguments explicitly.
}
