% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-files.R
\name{split_roxygen_segments}
\alias{split_roxygen_segments}
\title{Split text into roxygen and non-roxygen example segments}
\usage{
split_roxygen_segments(text, roxygen_examples)
}
\arguments{
\item{text}{Roxygen comments}

\item{roxygen_examples}{Integer sequence that indicates which lines in \code{text}
are roxygen examples. Most conveniently obtained with
\link{identify_start_to_stop_of_roxygen_examples_from_text}.}
}
\value{
A list with two elements:
\itemize{
\item A list that contains elements grouped into roxygen and non-roxygen
sections. This list is named \code{separated}.
\item An integer vector with the indices that correspond to roxygen code
examples in \code{separated}.
}
}
\description{
Split text into roxygen and non-roxygen example segments
}
\keyword{internal}
