% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{start_on_line}
\alias{start_on_line}
\title{Get the start right}
\usage{
start_on_line(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\value{
The line number on which the first token occurs.
}
\description{
On what line does the first token occur?
}
