% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_pd.R
\name{strip_eol_spaces}
\alias{strip_eol_spaces}
\title{Strip EOL spaces}
\usage{
strip_eol_spaces(pd_flat)
}
\arguments{
\item{pd_flat}{A flat parse table.}
}
\value{
A nested parse table.
}
\description{
Remove end-of-line spaces.
}
