% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_pd.R
\name{strip_eol_spaces_nested}
\alias{strip_eol_spaces_nested}
\title{Strip EOL spaces}
\usage{
strip_eol_spaces_nested(pd_nested)
}
\arguments{
\item{pd_nested}{A nested parse table.}
}
\value{
A nested parse table.
}
\description{
Remove end-of-line spaces.
}
