% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{style_active_pkg}
\alias{style_active_pkg}
\title{Wrapper around \code{\link[=style_pkg]{style_pkg()}} for access via Addin.}
\usage{
style_active_pkg()
}
\description{
Wrapper around \code{\link[=style_pkg]{style_pkg()}} for access via Addin.
}
\keyword{internal}
