% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{style_active_region}
\alias{style_active_region}
\title{Style the highlighted region}
\usage{
style_active_region()
}
\description{
Helper function for RStudio Addin. This function is complicated because of
one thing: You can highlight also just parts of lines.
}
