% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{style_book}
\alias{style_book}
\title{Prettify book source code}
\usage{
style_book(book = ".", ..., style = tidyverse_style,
  transformers = style(...), filetype = "Rmd", exclude_files = NULL,
  include_roxygen_examples = TRUE)
}
\arguments{
\item{book}{Path to a (subdirectory of an) book project.}

\item{...}{Arguments passed on to the \code{style} function.}

\item{style}{A function that creates a style guide to use, by default
\code{\link[=tidyverse_style]{tidyverse_style()}} (without the parentheses). Not used
further except to construct the argument \code{transformers}. See
\code{\link[=style_guides]{style_guides()}} for details.}

\item{transformers}{A set of transformer functions. This argument is most
conveniently constructed via the \code{style} argument and \code{...}. See
'Examples'.}

\item{filetype}{Vector of file extensions indicating which file types should
be styled. Case is ignored, and the \code{.} is optional, e.g. \code{c(".R", ".Rmd", ".Rnw")} or \code{c("r", "rmd", "rnw")}.}

\item{exclude_files}{Character vector with paths to files that should be
excluded from styling.}

\item{include_roxygen_examples}{Whether or not to style code in roxygen
examples.}
}
\description{
Performs various substitutions in all \code{.Rmd} files in a bookdown project.
One can also (optionally) style \code{.R} and/or \code{.Rnw} files in the project by
changing the \code{filetype} argument.
Carefully examine the results after running this function!
}
\examples{
\dontrun{

style_book()
style_book(filetype = c(".R", ".Rmd", ".Rnw"))
}
}
\seealso{
Other stylers: \code{\link{style_dir}},
  \code{\link{style_file}}, \code{\link{style_pkg}},
  \code{\link{style_text}}, \code{\link{styler_addins}}
}
\concept{stylers}
