% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{style_files}
\alias{style_files}
\title{Style files}
\usage{
style_files(path, ..., style = tidyverse_style, transformers = style(...))
}
\arguments{
\item{path}{A path to a file to style.}

\item{...}{Arguments passed on to the \code{style} function.}

\item{style}{A function that creates a style guide to use, by default
\code{\link[=tidyverse_style]{tidyverse_style()}} (without the parentheses). Not used
further except to construct the argument \code{transformers}. See
\code{\link[=style_guides]{style_guides()}} for details.}

\item{transformers}{A set of transformer functions. This argument is most
conveniently constructed via the \code{style} argument and \code{...}. See
'Examples'.}
}
\description{
Like \code{\link[=style_file]{style_file()}}, but can style more than one file at once.
}
\seealso{
Other stylers: \code{\link{style_dir}},
  \code{\link{style_file}}, \code{\link{style_pkg}},
  \code{\link{style_text}}, \code{\link{styler_addins}}
}
