% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_guides.R
\name{style_guides}
\alias{style_guides}
\title{Style guides}
\description{
Format code according to a style guide. Style guides are the input to the
argument \code{style} in \code{\link[=style_file]{style_file()}} and friends.
The available style guides are:
\itemize{
\item the tidyverse style guide (see \code{\link[=tidyverse_style]{tidyverse_style()}}).
}
}
