% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{style_region}
\alias{style_region}
\title{Style a region of text given context}
\usage{
style_region(text, context)
}
\arguments{
\item{text}{Character vector that contains the highlighted region.}

\item{context}{The context from \code{\link[=find_active_context]{find_active_context()}}.}
}
\description{
First extracts the relevant expression from \code{text} and returns after styling.
}
