% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-examples.R
\name{style_roxygen_code_example_one}
\alias{style_roxygen_code_example_one}
\title{Style a roxygen code example with exactly one \verb{@example} or \verb{@exampleIf}}
\usage{
style_roxygen_code_example_one(example_one, transformers, base_indention)
}
\arguments{
\item{example_one}{A character vector, one element per line, that contains in
total at most one example tag.}

\item{transformers}{Passed to \code{\link[=cache_make_key]{cache_make_key()}} to generate
a key.}

\item{base_indention}{Integer scalar indicating by how many spaces the whole
output text should be indented. Note that this is not the same as splitting
by line and add a \code{base_indention} spaces before the code in the case
multi-line strings are present. See 'Examples'.}
}
\description{
Style a roxygen code example with exactly one \verb{@example} or \verb{@exampleIf}
}
\keyword{internal}
