% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-roxygen-examples.R
\name{style_roxygen_code_examples}
\alias{style_roxygen_code_examples}
\title{TODO:
\itemize{
\item move to R/ui.R
\item same arguments as other stylers
\item add include_roxygen_code_examples argument to style_pkg() et al.
\item export
}}
\usage{
style_roxygen_code_examples(path)
}
\description{
TODO:
\itemize{
\item move to R/ui.R
\item same arguments as other stylers
\item add include_roxygen_code_examples argument to style_pkg() et al.
\item export
}
}
