% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-roxygen-examples.R
\name{style_roxygen_code_examples_one}
\alias{style_roxygen_code_examples_one}
\title{Style a roxygen code example that may contain a dontrun}
\usage{
style_roxygen_code_examples_one(example, transformers)
}
\arguments{
\item{example}{Roxygen example code.}
}
\description{
Parses roxygen2 comments into code, breaks it into dontrun / run sections and
processes each segment indicidually using
\code{\link[=style_roxygen_dontrun_code_examples_one]{style_roxygen_dontrun_code_examples_one()}}.
}
\keyword{internal}
