% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-roxygen-examples.R
\name{style_roxygen_dont_code_examples_one}
\alias{style_roxygen_dont_code_examples_one}
\title{Given a code segment is dont* or run, style it}
\usage{
style_roxygen_dont_code_examples_one(code_segment, transformers, is_dont)
}
\arguments{
\item{code_segment}{A character vector with code to style.}

\item{transformers}{A list of \emph{named} transformer functions}

\item{is_dont}{Whether the segment to process is a dontrun, dontshow,
donttest segemnt or not.}
}
\description{
Given a code segment is dont* or run, style it
}
\keyword{internal}
