% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{style_selection}
\alias{style_selection}
\title{Style the highlighted region}
\usage{
style_selection()
}
\description{
Helper function for RStudio Addin. This function is complicated because of
one thing: You can highlight also just parts of lines.
}
