% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-spaces.R
\name{style_space_around_math_token}
\alias{style_space_around_math_token}
\title{Style spacing around math tokens}
\usage{
style_space_around_math_token(strict, zero, one, pd_flat)
}
\arguments{
\item{strict}{Whether the rules should be applied strictly or not.}

\item{zero}{Character vector of tokens that should be surrounded with zero
spaces.}

\item{one}{Character vector with tokens that should be surrounded by at
least one space (depending on \code{strict = TRUE} in the styling functions
\code{\link[=style_text]{style_text()}} and friends). See 'Examples'.}

\item{pd_flat}{A nest or a flat parse table.}
}
\description{
Style spacing around math tokens
}
\keyword{internal}
