% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-spaces.R
\name{style_space_around_token}
\alias{style_space_around_token}
\title{Set spacing of token to a certain level}
\usage{
style_space_around_token(
  pd_flat,
  strict,
  tokens,
  level_before,
  level_after = level_before
)
}
\arguments{
\item{pd_flat}{A nest or a flat parse table.}

\item{strict}{Whether the rules should be applied strictly or not.}

\item{tokens}{Character vector with tokens that should be styled.}

\item{level_before, level_after}{Scalar indicating the amount of spaces that
should be inserted around the \code{tokens} on the left and right position
respectively.}
}
\description{
Set the spacing of all \code{tokens} in \code{pd_flat} to \code{level} if \code{strict = TRUE} or
to at least to \code{level} if \code{strict = FALSE}.
}
\keyword{internal}
