% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{style_src}
\alias{style_src}
\title{Prettify arbitrary R code}
\usage{
style_src(path = ".", transformers = get_transformers(), recursive = TRUE)
}
\arguments{
\item{path}{Path to a directory with files to transform.}

\item{transformers}{A list of transformer functions to be applied to the
files in \code{path}.}

\item{recursive}{A logical value indicating whether or not files in subdirectories
of \code{path} should be styled as well.}
}
\description{
Performs various substitutions in all \code{.R} files in a directory.
Carefully examine the results after running this function!
}
