% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styler.R
\docType{package}
\name{styler-package}
\alias{styler}
\alias{styler-package}
\title{Non-invasive pretty printing of R code}
\description{
styler allows you to format .R files, packages or entire R source trees
according to a style guide. See the INDEX for more information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/styler}
  \item \url{http://krlmlr.github.io/styler}
  \item Report bugs at \url{https://github.com/krlmlr/styler/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}
