% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{styler}
\alias{styler}
\title{Prettify R source code}
\usage{
styler(pkg = ".")
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package}
}
\description{
Performs various substitutions in all \code{.R} files in a package.
Carefully examine the results after running this function!
}
