% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.R
\name{styler_future}
\alias{styler_future}
\title{Parallelization of styling}
\description{
\code{\link[=style_file]{style_file()}}, \code{\link[=style_pkg]{style_pkg()}} and \code{\link[=style_dir]{style_dir()}} leverage the \code{{future}}
framework for parallelization. To make use of parallel processing, you need
to have the package \code{{furrr}} installed. In that case, the strategy
\code{\link[future:multiprocess]{future::multiprocess()}} is set temporarily during styling (only if there is
more than two files to style, because of start-up costs of parallelization).
You can prevent this temporary change in the future strategy by setting the
environment variable \code{R_STYLER_FUTURE_NO_OVERRIDE} to \code{TRUE} (case ignored). In
that case, the future strategy set before calling styler will be used, if
you have not set any, the future \code{{framework}} falls back to
\code{\link[future:sequential]{future::sequential()}}.
}
\section{Life-cycle}{

The parallelization feature is experimental, in particular how its governed
with \code{R_STYLER_FUTURE_NO_OVERRIDE} and how progress bars are displayed with
\code{{furrr}}. In the future, most likely the \code{{progressr}} package will be used
to handle progress bars.
}

