% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_collection}
\alias{test_collection}
\title{Run a collection of tests}
\usage{
test_collection(
  test,
  sub_test = NULL,
  dry = "off",
  write_tree = FALSE,
  transformer,
  ...
)
}
\arguments{
\item{test}{The test to run. It corresponds to a folder name in
tests/testthat.}

\item{sub_test}{A regex pattern to further reduce the amount of test files
to be tested in the test. \code{sub_test} must match the beginning of file
names in tests/testthat. \code{NULL} matches all files.}

\item{dry}{To indicate whether styler should run in \emph{dry} mode, i.e. refrain
from writing back to files. Neither \code{"on"} nor \code{"fail"} write back; the
latter returns an error if the input code is not identical to the result
of styling. "off", the default, writes back if the input and output of
styling are not identical.}

\item{write_tree}{Whether or not the tree structure of the test should be
computed and written to a files.}

\item{transformer}{A function to apply to the content of \code{in_item}.}

\item{...}{Parameters passed to transformer function.}
}
\description{
Run transformations on all *-in.R files in a test directory and compare them
with their *-out.R counterpart.
}
\details{
Each file name that matches \code{test} and \code{sub_test} and ends with
"-in.R" is considered as an input to test. Its counterpart,
the reference to compare it against is the *-out.R file. It is constructed
by taking the substring of the *-in.R file before the
last dash and adding -out.R. In contrast to older versions of this
function, every *-out.R file has just one in file.
}
\keyword{internal}
