% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing-public-api.R
\name{test_dry}
\alias{test_dry}
\title{Test the dry argument}
\usage{
test_dry(path, styler, styled = FALSE)
}
\arguments{
\item{path}{A path to pass to the \code{styler}.}

\item{styler}{A function that takes \code{path}, typically a user exposed styler
function that has side effects, like \code{\link[=style_file]{style_file()}}.}
}
\description{
Test the dry argument
}
\keyword{internal}
