% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{testthat_file}
\alias{testthat_file}
\title{Create the path to a test that file}
\usage{
testthat_file(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=file.path]{file.path()}} to construct the path after
".../tests/testthat/"}
}
\description{
Create the path to a test that file
}
\keyword{internal}
