% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_guides.R
\name{tidyverse_style}
\alias{tidyverse_style}
\title{The tidyverse style}
\usage{
tidyverse_style(scope = "tokens", strict = TRUE, indent_by = 2,
  start_comments_with_one_space = FALSE)
}
\arguments{
\item{scope}{The extent of manipulation. Can range from "none" (least
invasive) to "token" (most invasive). See 'Details'. This argument is a
vector of length one.}

\item{strict}{A logical value indicating whether a set of strict
or not so strict transformer functions should be returned.}

\item{indent_by}{How many spaces of indention should be inserted after
operators such as '('.}

\item{start_comments_with_one_space}{Whether or not comments should start
with only one space (see \code{\link[=start_comments_with_space]{start_comments_with_space()}}).}
}
\description{
The tidyverse style
}
\details{
The following options for \code{scope} are available.
\itemize{
\item "none": Performs no transformation at all.
\item "spaces": Manipulates spacing between token on the same line.
\item "indention": In addition to "spaces", this option also manipulates the
indention level.
\item "line_breaks": In addition to "indention", this option also manipulates
line breaks.
\item "tokens": In addition to "line_breaks", this option also manipulates
tokens.
}

As it becomes clear from this description, more invasive operations can only
be performed if all less invasive operations are performed too.
}
