% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.R
\name{token_is_adjacent}
\alias{token_is_adjacent}
\title{Are they neighbours?}
\usage{
token_is_adjacent(pd, token)
}
\arguments{
\item{pd}{A parse table.}

\item{token}{A vector of tokens to check.}
}
\description{
Check whether a token is positioned next to another token in a parse table.
}
