% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-alignment.R
\name{token_is_on_alligned_line}
\alias{token_is_on_alligned_line}
\title{Check if tokens are aligned}
\usage{
token_is_on_alligned_line(pd_flat, op_before)
}
\description{
A line is called aligned if the following conditions hold:
}
\details{
\itemize{
\item argument name and = have the same line1 (i.e. the start of a token) for
multiple lines for the first column.
\item spacing around comma is correct (none before, at least one after).
\item and at least one space around =.
}
}
