% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{token_on_same_line_as}
\alias{token_on_same_line_as}
\title{Return the indices of a token on the line of a target index}
\usage{
token_on_same_line_as(flattened_pd, target_index, token)
}
\description{
Returns only tokens that are coming \emph{after} target_index and are on
the same line.
}
