% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.R
\name{tokenize}
\alias{tokenize}
\title{Obtain token table from text}
\usage{
tokenize(text)
}
\arguments{
\item{text}{A character vector.}
}
\value{
A flat parse table
}
\description{
\code{\link[utils:getParseData]{utils::getParseData()}} is used to obtain a flat parse table from \code{text}.
}
