% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{transform_lines_utf8}
\alias{transform_lines_utf8}
\title{Apply a function to the contents of a file}
\usage{
transform_lines_utf8(path, fun, write_back = TRUE)
}
\arguments{
\item{path}{A vector with file paths to transform.}

\item{fun}{A function that returns a character vector.}

\item{write_back}{Whether or not the results of the transformation should
be written back to the file.}
}
\description{
Transforms a file with a function.
}
\keyword{internal}
