% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-code.R
\name{transform_mixed}
\alias{transform_mixed}
\title{Transform mixed contents}
\usage{
transform_mixed(lines, transformer_fun, filetype)
}
\arguments{
\item{lines}{A character vector of lines from an Rmd or Rnw file.}

\item{transformer_fun}{A styler transformer function.}

\item{filetype}{A string indicating the filetype - either 'Rmd' or 'Rnw'.}
}
\description{
Applies the supplied transformer function to code chunks identified within
an Rmd or Rnw file and recombines the resulting (styled) code chunks with the text
chunks.
}
\keyword{internal}
