% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-code.R
\name{transform_mixed_non_empty}
\alias{transform_mixed_non_empty}
\title{Ensure for \code{.Rmd} and friends that a code chunk without code is formatted as
a code chunk without any lines.}
\usage{
transform_mixed_non_empty(r_chunk, transformer_fun)
}
\description{
Ensure for \code{.Rmd} and friends that a code chunk without code is formatted as
a code chunk without any lines.
}
\keyword{internal}
