% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_nested}
\alias{transform_nested}
\title{Transform nested tibble with a transformer function}
\usage{
transform_nested(pd_nested, fun)
}
\arguments{
\item{pd_nested}{A nested parse table.}

\item{fun}{A transformer function that operates on a flat parse table.}
}
\value{
A transformed, nested parse table. Whereas \code{fun} was applied to
every level of nesting within \code{pd_nested}.
}
\description{
Transform nested tibble with a transformer function
}
