% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-code.R
\name{transform_rnw}
\alias{transform_rnw}
\title{Transform Rnw contents}
\usage{
transform_rnw(lines, transformer_fun)
}
\arguments{
\item{lines}{A character vector of lines from an Rnw file}

\item{transformer_fun}{A styler transformer function}
}
\description{
Applies the supplied transformer function to code chunks identified within
an Rnw file and recombines the resulting (styled) code chunks with the text
chunks.
}
\keyword{internal}
