% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{transform_utf8}
\alias{transform_utf8}
\title{Apply a function to the contents of a file}
\usage{
transform_utf8(path, fun, dry)
}
\arguments{
\item{path}{A vector with file paths to transform.}

\item{fun}{A function that returns a character vector.}

\item{dry}{To indicate whether styler should run in \emph{dry} mode, i.e. refrain
from writing back to files. Neither \code{"on"} nor \code{"fail"} write back; the
latter returns an error if the input code is not identical to the result
of styling. "off", the default, writes back if the input and output of
styling are not identical.}
}
\description{
Transforms a file with a function.
}
\keyword{internal}
