% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-files.R
\name{transformers_subset}
\alias{transformers_subset}
\title{Remove transformers that are not needed
For every transformer, at least one token must be given to make subsetting.
active.}
\usage{
transformers_subset(text, transformers)
}
\description{
Remove transformers that are not needed
For every transformer, at least one token must be given to make subsetting.
active.
}
