% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-files.R
\name{transformers_subset_impl}
\alias{transformers_subset_impl}
\title{Removes the transformers if tokens in x}
\usage{
transformers_subset_impl(transformers, token, scope, code)
}
\arguments{
\item{transformers}{The style full style guide, i.e. the result of
\code{\link[=create_style_guide]{create_style_guide()}}.}

\item{token}{Named character vector: Names are rules and values are the token
that trigger are required to be absent to trigger a removal.}

\item{scope}{The low-level scope, e.g. 'token'.}

\item{code}{tokenized code for which we check if \code{token} is in them.}
}
\description{
Removes the transformers if tokens in x
}
