% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialized_tests.R
\name{tree_from_in}
\alias{tree_from_in}
\title{Construct paths of a tree object given the paths of *-in.R files}
\usage{
tree_from_in(in_files, suffix = "_tree")
}
\arguments{
\item{in_files}{Character vector of *-in.R files.}

\item{suffix}{Suffix for the tree object.}
}
\description{
Construct paths of a tree object given the paths of *-in.R files
}
