% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{try_transform_as_r_file}
\alias{try_transform_as_r_file}
\title{Style a file as if it was an .R file}
\usage{
try_transform_as_r_file(context, transformer)
}
\arguments{
\item{context}{The context from \code{styler:::get_rstudio_context()}.}

\item{transformer}{A transformer function most conveniently constructed with
\code{\link[=make_transformer]{make_transformer()}}.}
}
\description{
If not successful, the file is most
likely not a .R file, so saving the file and try styling again will work if
the file is an .Rmd file. Otherwise, we can throw an error that the file must
be a .R or .Rmd file.
}
