% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{two_cols_match}
\alias{two_cols_match}
\title{Check whether two columns match}
\usage{
two_cols_match(col1, col2, data)
}
\arguments{
\item{col1, col2}{Column names as string.}

\item{data}{The data frames that contains \code{col1} and \code{col2}.}
}
\description{
Check whether two columns match
}
