% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unindent.R
\name{unindent_child}
\alias{unindent_child}
\title{Unindent a child}
\usage{
unindent_child(pd, token = c("')'", "'}'"), unindent_by = 2L)
}
\arguments{
\item{pd}{A parse table.}

\item{token}{The token the unindention should be based on.}

\item{unindent_by}{By how many spaces one level of indention is reversed.}
}
\description{
Unindent a child
}
\keyword{internal}
