% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-indention.R
\name{unindent_fun_dec}
\alias{unindent_fun_dec}
\title{Revert the indention of function declaration header}
\usage{
unindent_fun_dec(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
Necessary for consistent indention of the function declaration header.
}
\seealso{
set_unindention_child update_indention_ref_fun_dec
}
\keyword{internal}
