% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{uninstall_precommit}
\alias{uninstall_precommit}
\title{Unistall pre-commit}
\usage{
uninstall_precommit(scope = "repo", ask = "global", path_root = ".")
}
\arguments{
\item{scope}{Either "repo" or "global". "repo" removes pre-commit from your
project, but you will be able to use it in other projects. With "global",
you remove the pre-commit executable in the virtual python environment
r-reticulate so it won't be available in any project.}
}
\description{
Unistall pre-commit
}
