% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_pd.R
\name{update_indention_nested}
\alias{update_indention_nested}
\alias{indent_round_nested}
\title{Update indention information of nested parse data}
\usage{
indent_round_nested(pd)
}
\arguments{
\item{pd}{A nested parse table that is already enhanced with
line break and space information via \link{create_filler_nested}.}
}
\description{
These functions apply the update functions of the same name but without
suffix nested to each level of nesting of the nested parse table.
}
