% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_precommit}
\alias{update_precommit}
\title{Update the pre-commit executable}
\usage{
update_precommit()
}
\description{
Updates the conda installation of the upstream framework pre-commit. This
does not update the R package \code{{precommit}}. it only works if you
framework.
}
\seealso{
Other executable managers: 
\code{\link{version_precommit}()}
}
\concept{executable managers}
