% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{update_rev_in_config}
\alias{update_rev_in_config}
\title{Updates the hook version ref of {precommit} in a \code{.pre-commit-config} file}
\usage{
update_rev_in_config(new_version, path = "inst/pre-commit-config.yaml")
}
\arguments{
\item{new_version}{The version string of the new version.}

\item{path}{The path to a pre-commit config file.}
}
\description{
This is useful in the release process because when releasing a new version,
we must make sure the template that is used with \code{precommit::use_precommit()}
is up-to date. Also, after we pushed the release to GitHub, we want to update
the hooks from our own hook repo in the source repo too (we could also do that
with \code{precommit::autoupdate()} though).
}
\keyword{internal}
