% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_precommit}
\alias{use_precommit}
\title{Set up a pre-com
Copy a default pre-commit-config.yaml in the root of your project}
\usage{
use_precommit(hooks = NULL, path_root = here::here(), force = FALSE)
}
\arguments{
\item{hooks}{Ignored.}

\item{path_root}{The path to the root directory of your project.}
}
\description{
Set up a pre-com
Copy a default pre-commit-config.yaml in the root of your project
}
