% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{use_precommit_config}
\alias{use_precommit_config}
\title{Initiate a pre-commit config file}
\usage{
use_precommit_config(
  path_cp_config_from = getOption("precommit.path_cp_config_from"),
  force,
  open = rstudioapi::isAvailable(),
  path_root = here::here()
)
}
\arguments{
\item{path_cp_config_from}{Path or URL to a \code{.pre-commit-config.yaml}. This config file
will be hard-copied into \code{path_root}. If \code{NULL}, we use
a default config from the path returned by
\code{system.file("pre-commit-config.yaml", package = "precommit")}. See
section 'Copying an existing config file'.}

\item{force}{Whether to replace an existing config file.}

\item{open}{Whether or not to open the .pre-commit-config.yaml after
it's been placed in your repo. The default is \code{TRUE} when working in
RStudio. Otherwise, we recommend manually inspecting the file.}

\item{path_root}{The path to the root directory of your project.}
}
\description{
Initiate a pre-commit config file
}
\section{Copying an existing config file}{

You can use an existing \code{.pre-commit-config.yaml} file when intiializing
pre-commit with \code{\link[=use_precommit]{use_precommit()}} using the argument \code{path_cp_config_from} to copy
an existing config file into your repo. For convenience, this argument
defaults to the R option \code{precommit.path_cp_config_from}, so you may want to
set this option in your \code{.Rprofile} for convenience.
Note that this is \strong{not} equivalent
to the \code{--config} option in the CLI command \verb{pre-commit install} and similar,
which do \emph{not} copy a config file into a project root (and allow to put it
under version control), but rather link it in some more or less transparent
way.
}

