% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-create.R
\name{validate_new_pos_id}
\alias{validate_new_pos_id}
\title{Validate new position ids}
\usage{
validate_new_pos_id(new_ids, after)
}
\arguments{
\item{new_ids}{A vector with new ids}

\item{after}{Whether the ids are created with \code{after = TRUE} (and hence
should be in the range x.0-x.45) or not.}
}
\description{
Validate new position ids
}
\seealso{
Other token creators: \code{\link{create_pos_id}},
  \code{\link{create_tokens}}
}
