% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-create.R
\name{validate_new_pos_ids}
\alias{validate_new_pos_ids}
\title{Validate sequence of new position ids}
\usage{
validate_new_pos_ids(new_ids, after)
}
\arguments{
\item{new_ids}{A vector with new ids}

\item{after}{Whether the ids are created with \code{after = TRUE} (and hence
should be in the range x.0-x.45) or not.}
}
\description{
Ids created with \code{after = TRUE} can have \code{pos_id} values between x.0 and
x.5 and ids created with \code{after = FALSE} can have \code{pos_id} values between
1+ x.0 and 1 + x.5 where x is the \code{pos_id} integer which was used as a
reference to create the new \code{pos_ids}.
}
\seealso{
Other token creators: 
\code{\link{create_pos_ids}()},
\code{\link{create_tokens}()}
}
\concept{token creators}
\keyword{internal}
