% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-files.R
\name{verify_roundtrip}
\alias{verify_roundtrip}
\title{Verify the styling if possible}
\usage{
verify_roundtrip(old_text, new_text, transformers)
}
\arguments{
\item{old_text}{The initial expression in its character representation.}

\item{new_text}{The styled expression in its character representation.}

\item{transformers}{The list of transformer functions used for styling.
Needed for reverse engineering the scope.}
}
\description{
If scope was set to "line_breaks" or lower (compare \code{\link[=tidyverse_style]{tidyverse_style()}}),
we can compare the expression before and after styling and return an error if
it is not the same. Note that this method ignores comments and no
verification can be conducted if scope > "line_breaks".
}
\examples{
styler:::verify_roundtrip("a+1", "a + 1", tidyverse_style(scope = "line_breaks"))
styler:::verify_roundtrip(
  "a+1",
  "a + 1 # comments are dropped",
  tidyverse_style(scope = "line_breaks")
)
\dontrun{
styler:::verify_roundtrip("a+1", "b - 3", tidyverse_style(scope = "line_breaks"))
}
}
