% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{version_precommit}
\alias{version_precommit}
\title{Retrieve the version of the pre-commit executable used}
\usage{
version_precommit()
}
\description{
Retrieves the version of the pre-commit executable used.
}
\seealso{
Other executable managers: 
\code{\link{install_precommit}()},
\code{\link{uninstall_precommit}()},
\code{\link{update_precommit}()}
}
\concept{executable managers}
