% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{visit}
\alias{visit}
\title{Visit'em all}
\usage{
visit(pd_nested, funs)
}
\arguments{
\item{pd_nested}{A nested parse table.}

\item{funs}{A list of transformer functions.}
}
\description{
Apply a list of functions to each level in a nested parse table.
}
\seealso{
Other visitors: \code{\link{visit_one}}
}
