% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit.R
\name{visit_one}
\alias{visit_one}
\title{Transform a flat parse table with a list of transformers}
\usage{
visit_one(pd_flat, funs)
}
\arguments{
\item{pd_flat}{A flat parse table.}

\item{funs}{A list of transformer functions.}
}
\description{
Uses \code{\link[=Reduce]{Reduce()}} to apply each function of \code{funs} sequentially to
\code{pd_flat}.
}
\seealso{
Other visitors: 
\code{\link{visit}}
}
\concept{visitors}
\keyword{internal}
