% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-other.R
\name{wrap_else_multiline_curly}
\alias{wrap_else_multiline_curly}
\title{Add curly braces to else}
\usage{
wrap_else_multiline_curly(pd, indent_by = 2, space_after = 0)
}
\arguments{
\item{pd}{A parse table.}

\item{indent_by}{The amount of spaces used to indent an expression in curly
braces. Used for unindention.}

\item{space_after}{How many spaces should be inserted after the closing brace.}
}
\description{
Wrap the else part of a conditional expression into curly braces if not
already wrapped into a such.
}
\keyword{internal}
