% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.R
\name{wrap_expr_in_expr}
\alias{wrap_expr_in_expr}
\title{Wrap an expression into an expression}
\usage{
wrap_expr_in_expr(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
Takes a parse table and wraps it in a new parse table that contains the
expression as a child.
}
