% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-tokens.R
\name{wrap_if_else_while_for_fun_multi_line_in_curly}
\alias{wrap_if_else_while_for_fun_multi_line_in_curly}
\title{Wrap if-else, while and for statements in curly braces}
\usage{
wrap_if_else_while_for_fun_multi_line_in_curly(pd, indent_by = 2L)
}
\arguments{
\item{pd}{A parse table.}

\item{indent_by}{The amount of spaces used to indent an expression in curly
braces. Used for unindention.}
}
\description{
Wrap statements in curly braces if it is not already wrapped in a such.
}
\keyword{internal}
