% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-tokens.R
\name{wrap_multiline_curly}
\alias{wrap_multiline_curly}
\title{Wrap a multi-line statement in curly braces}
\usage{
wrap_multiline_curly(pd, indent_by, key_token, space_after = 1L)
}
\arguments{
\item{pd}{A parse table.}

\item{indent_by}{The amount of spaces used to indent an expression in curly
braces. Used for unindention.}

\item{key_token}{The token that comes right before the token that contains
the expression to be wrapped (ignoring comments). For if and while loops,
this is the closing "')'", for a for-loop it's "forcond".}

\item{space_after}{How many spaces should be inserted after the closing brace.}
}
\description{
Wrap a multi-line statement in curly braces
}
\keyword{internal}
