% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_utf8}
\alias{write_utf8}
\title{Drop-in replacement for \code{xfun::write_utf8()}}
\usage{
write_utf8(text, con, ...)
}
\description{
Drop-in replacement for \code{xfun::write_utf8()}
}
\keyword{internal}
