% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-guides.R
\name{yihui_style}
\alias{yihui_style}
\title{The yihui style}
\usage{
yihui_style()
}
\description{
Style code according to \href{https://github.com/yihui}{Yihui Xie}.
}
\details{
This style guide is the same as \code{\link[=tidyverse_style]{tidyverse_style()}}, except it uses
equals (\code{=}) rather than arrow \code{<-} assignment.
}
\examples{
style_text("x <- 1", style = yihui_style)
}
\seealso{
Other style_guides: \code{\link{tidyverse_style}}
}
\concept{style_guides}
