% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-examples-add-remove.R
\name{add_roxygen_mask}
\alias{add_roxygen_mask}
\title{Add the roxygen mask to code}
\usage{
add_roxygen_mask(text, example_type)
}
\arguments{
\item{text}{Character vector with code.}

\item{example_type}{Either 'examples' or 'examplesIf'.}
}
\description{
Add the roxygen mask to code
}
\keyword{internal}
