% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.R
\name{add_token_terminal}
\alias{add_token_terminal}
\alias{add_terminal_token_after}
\alias{add_terminal_token_before}
\title{Add information about previous / next token to each terminal}
\usage{
add_terminal_token_after(pd_flat)

add_terminal_token_before(pd_flat)
}
\arguments{
\item{pd_flat}{A flat parse table.}
}
\description{
Add information about previous / next token to each terminal
}
