% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{cache_activate}
\alias{cache_activate}
\alias{cache_deactivate}
\title{Activate or deactivate the styler cache}
\usage{
cache_activate(cache_subdir = NULL)

cache_deactivate()
}
\arguments{
\item{cache_subdir}{The sub-directory you want to use under the R.cache
location. If \code{NULL}, the option "styler.cache_subdir" is considered
which defaults to the version of styler used.}
}
\description{
Helper functions to control the behavior of caching. Simple wrappers around
\code{\link[base:options]{base::options()}}.
}
\seealso{
Other cache managers: \code{\link{cache_clear}},
  \code{\link{cache_info}}
}
\concept{cache managers}
