% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-caching.R
\name{cache_activate}
\alias{cache_activate}
\alias{cache_deactivate}
\title{Activate or deactivate the styler cache}
\usage{
cache_activate(cache_name = NULL, verbose = TRUE)

cache_deactivate(verbose = TRUE)
}
\arguments{
\item{cache_name}{The name of the styler cache to use. If
\code{NULL}, the option "styler.cache_name" is considered which defaults to
the version of styler used.}

\item{verbose}{Whether or not to print an informative message about what the
function is doing.}
}
\description{
Helper functions to control the behavior of caching. Simple wrappers around
\code{\link[base:options]{base::options()}}.
}
\seealso{
Other cache managers: 
\code{\link{cache_clear}()},
\code{\link{cache_info}()}
}
\concept{cache managers}
