% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{cache_by_expression}
\alias{cache_by_expression}
\title{Cache text}
\usage{
cache_by_expression(text, transformers)
}
\arguments{
\item{text}{A character vector with one or more expressions.}

\item{transformers}{The transformers.}
}
\description{
Splits \code{text} into expressions and adds these to the cache. Note that
comments are \strong{not} cached because caching them is too expensive.
}
\keyword{internal}
