% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{cache_clear}
\alias{cache_clear}
\title{Clear the cache}
\usage{
cache_clear(cache_name = NULL, ask = TRUE)
}
\arguments{
\item{cache_name}{The name of the cache to use. If
\code{NULL}, the option "styler.cache_name" is considered which defaults to
the version of styler used.}

\item{ask}{Whether or not to interactively ask the user again.}
}
\description{
Clears the cache that stores which files' styling is up-to-date. You won't be
able to undo this. Note that the file corresponding to the cache (a folder
on our file stystem) won't be deleted, but it will be empty after calling
\code{cache_clear}.
If the cache is used at all or not is determined via the R option
\code{styler.use_cache}, defaulting to \code{TRUE}.
}
\details{
Each version of styler has it's own cache by default, because styling is
potentially different with different versions of styler.
}
\seealso{
Other cache managers: \code{\link{cache_activate}},
  \code{\link{cache_info}}
}
\concept{cache managers}
