% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{cache_find_path}
\alias{cache_find_path}
\title{Where is the cache?}
\usage{
cache_find_path(cache_name = NULL)
}
\arguments{
\item{cache_name}{The name of the styler cache to use. If
\code{NULL}, the option "styler.cache_name" is considered which defaults to
the version of styler used.}
}
\description{
Finds the path to the cache and creates it if it does not exist.
}
\keyword{internal}
