% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{cache_info}
\alias{cache_info}
\title{Show information about the styler cache}
\usage{
cache_info(cache_subdir = NULL, format = "lucid")
}
\arguments{
\item{format}{Either "lucid" for a printed summary or "tabular" for a
tabular summary.}
}
\description{
If the cache is used at all or not is determined via the R option
\code{styler.use_cache}.
}
