% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-caching.R
\name{caching}
\alias{caching}
\title{Remember the past to be quicker in the future}
\description{
styler by default uses caching. It may prompt you to install the R package
\code{R.cache} the first time you want to use it. R.cache will also ask you to let
it create a permanent cache on your file system that styler will use.
This is needed if you want to cache across R sessions and not just within.
The cache is specific to a version of styler by default, because different
versions potentially format code differently. This means after upgrading
styler or a style guide you use, the cache will be re-built.
See \code{\link[=cache_info]{cache_info()}},
\code{\link[=cache_activate]{cache_activate()}}, \code{\link[=cache_clear]{cache_clear()}} for utilities to manage the cache.
}
