% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{call_precommit}
\alias{call_precommit}
\title{Call pre-commit}
\usage{
call_precommit(...)
}
\arguments{
\item{...}{Arguments passed to the command line call \code{pre-commit}.}
}
\description{
Either via \verb{conda run} (because conda env needs to be activated in general to
ensure an executable to runs successfully) or, if the installation method was
not conda, as a plain bash command.
}
\keyword{internal}
