% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{compute_parent_to_update}
\alias{compute_parent_to_update}
\title{Compute the indices of the parents that are to be updated}
\usage{
compute_parent_to_update(flattened_pd)
}
\arguments{
\item{flattened_pd}{A flattened parse table.}
}
\description{
Compute the indices of the parents that are to be updated
}
\details{
Updating indention is carried out in two steps: updating the indention itself
and updating the position of the corresponding tokens in the colum \code{col}, so
subsequent indention based on re-indented tokens can be done correctly.

The idea is to first find the tokens on which the indention of other tokens
should be based on. Then, via \code{id}, it is possible to identify the closing
parenthesis. Since the column \code{indent} was removed in \code{\link[=enrich_terminals]{enrich_terminals()}}
and all indention information is now stored in the column \code{spaces} of the
first token of every line, spacing is only updated for the tokens after a
line break. Therefore, \code{lag_spaces} of these tokens will be updated with
\code{col} of the opening bracket.

All re-indented lines now have oudated \code{col} values. Hence, these now need to
be updated according to the number of spaces that were added
to or removed from the first token on the line. This is necessary because
subsequent re-indention depends on \code{col} values.
}
