% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.R
\name{compute_parse_data_nested}
\alias{compute_parse_data_nested}
\title{TODO:
\itemize{
\item Implement add_ws_to_parse_data_nested()
\itemize{
\item Walk tree defined by \code{child}, compute whitespace information
\item Store indention depth in a separate column, unaffected by
inter-token space
}
\item Implement compute_parse_data_nested_with_ws() as
compute_parse_data_nested() + add_ws_to_parse_data_nested()
\item Implement serialization of nested parse data
\item Use compute_parse_data_nested_with_ws() instead of
compute_parse_data_flat_with_ws()
\item Perform all transformations on hierarchical structure
\itemize{
\item Compute text for a sub-element
}
\item Compute indentation
\itemize{
\item Braces
\item Function calls
\item Function definitions
}
\item Remove \code{includeText = TRUE}
}}
\usage{
compute_parse_data_nested(text)
}
\description{
TODO:
\itemize{
\item Implement add_ws_to_parse_data_nested()
\itemize{
\item Walk tree defined by \code{child}, compute whitespace information
\item Store indention depth in a separate column, unaffected by
inter-token space
}
\item Implement compute_parse_data_nested_with_ws() as
compute_parse_data_nested() + add_ws_to_parse_data_nested()
\item Implement serialization of nested parse data
\item Use compute_parse_data_nested_with_ws() instead of
compute_parse_data_flat_with_ws()
\item Perform all transformations on hierarchical structure
\itemize{
\item Compute text for a sub-element
}
\item Compute indentation
\itemize{
\item Braces
\item Function calls
\item Function definitions
}
\item Remove \code{includeText = TRUE}
}
}
