% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{compute_parse_data_nested}
\alias{compute_parse_data_nested}
\title{Obtain a nested parse table from a character vector}
\usage{
compute_parse_data_nested(
  text,
  transformers = tidyverse_style(),
  more_specs = NULL
)
}
\arguments{
\item{text}{The text to parse.}

\item{transformers}{Passed to \code{\link[=cache_make_key]{cache_make_key()}} to generate a key.}

\item{more_specs}{Passed to \code{\link[=cache_make_key]{cache_make_key()}} to generate a key.}
}
\value{
A nested parse table. See \code{\link[=tokenize]{tokenize()}} for details on the columns
of the parse table.
}
\description{
Parses \code{text} to a flat parse table and subsequently changes its
representation into a nested parse table with \code{\link[=nest_parse_data]{nest_parse_data()}}.
}
\examples{
code <- "
ab     <- 1L # some comment
abcdef <- 2L
"
writeLines(code)
compute_parse_data_nested(code)
}
