% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{compute_parse_data_nested}
\alias{compute_parse_data_nested}
\title{Obtain a nested parse table from a character vector}
\usage{
compute_parse_data_nested(text, transformers)
}
\arguments{
\item{text}{A character vector to parse.}
}
\value{
A nested parse table. See \code{\link[=tokenize]{tokenize()}} for details on the columns
of the parse table.
}
\description{
Parses \code{text} to a flat parse table and subsequently changes its
representation into a nested parse table with \code{\link[=nest_parse_data]{nest_parse_data()}}.
}
\keyword{internal}
