% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{construct_out}
\alias{construct_out}
\title{Construct *-out.R from a *-in.R}
\usage{
construct_out(in_paths)
}
\arguments{
\item{in_paths}{A character vector that denotes paths to *-in.R files.}
}
\description{
Multiple *-in.R files can have the same *-out.R file since to create the
*-out.R file, everything after the first dash is replaced by *-out.R.
}
\examples{
styler:::construct_out(c(
  "path/to/file/first-in.R",
  "path/to/file/first-extended-in.R"
))
}
\keyword{internal}
