% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_newlines_to_linebreaks}
\alias{convert_newlines_to_linebreaks}
\title{Replace the newline character with a line break}
\usage{
convert_newlines_to_linebreaks(text)
}
\arguments{
\item{text}{A character vector}
}
\description{
Replace the newline character with a line break
}
\examples{
styler:::convert_newlines_to_linebreaks("x\n2")
# a simple strsplit approach does not cover both cases
unlist(strsplit("x\n\n2", "\n", fixed = TRUE))
unlist(strsplit(c("x", "", "2"), "\n", fixed = TRUE))
styler:::convert_newlines_to_linebreaks(c("x", "2"))
}
\keyword{internal}
