% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsed.R
\name{create_filler}
\alias{create_filler}
\title{Enrich parse table with space and linebreak information}
\usage{
create_filler(pd_flat)
}
\arguments{
\item{pd_flat}{A parse table.}
}
\value{
A parse table with two three columns: lag_newlines, newlines and
spaces.
}
\description{
This function computes difference (as column and line difference) between two
entries in the parse table and adds this information to the table.
}
\seealso{
\code{\link[=create_filler_nested]{create_filler_nested()}}
}
