% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-create.R
\name{create_pos_id}
\alias{create_pos_id}
\alias{create_pos_id_one}
\title{Create a valid pos_id if possible}
\usage{
create_pos_id(pd, pos, by = 0.1, after = FALSE, n = 1)

create_pos_id_one(pd, pos, after = FALSE, by = 0.1)
}
\arguments{
\item{pd}{A parse table.}

\item{pos}{The position where the new id should be inserted.}

\item{by}{By how much the reference \code{pos_id} should be increased / decreased
to create a new id.}

\item{after}{Boolean indicating whether it should be inserted after or before
\code{pos}.}

\item{n}{Number of ids to generate.}
}
\value{
Returns a valid pos_id or an error if it was not possible to create one. The
validation is done with \code{\link[=validate_new_pos_id]{validate_new_pos_id()}}
}
\description{
Create a valid pos_id if possible
}
\section{Functions}{
\itemize{
\item \code{create_pos_id_one}: Helper to create one id.
}}

\seealso{
Other token creators: \code{\link{create_tokens}},
  \code{\link{validate_new_pos_id}}
}
