% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{default_style_guide_attributes}
\alias{default_style_guide_attributes}
\title{Initialize default style guide attributes}
\usage{
default_style_guide_attributes(pd_flat)
}
\arguments{
\item{pd_flat}{A parse table.}
}
\description{
This function initializes and removes various variables from the parse
table.
}
\examples{
withr::with_options(
  list(styler.cache_name = NULL), # temporarily deactivate cache
  {
    string_to_format <- "call( 3)"
    pd <- styler:::compute_parse_data_nested(string_to_format)
    styler:::pre_visit(pd, c(default_style_guide_attributes))
  }
)
}
\keyword{internal}
