% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{diff_requires_run_roxygenize}
\alias{diff_requires_run_roxygenize}
\title{Check if we should run roxygen.}
\usage{
diff_requires_run_roxygenize(root = here::here())
}
\arguments{
\item{root}{The root of project.}
}
\value{
A logical vector of length 1.
}
\description{
This is the case if a new or replaced/removed line contains a roxygen2
comment in a file that is staged. This is only exported because it's used
in a hook script: It's not intended to be called by users of {precommit}
directly.
}
\examples{
\dontrun{
diff_requires_run_roxygenize()
}
}
