% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{diff_requires_run_roxygenize}
\alias{diff_requires_run_roxygenize}
\title{Check if we should run roxygen.}
\usage{
diff_requires_run_roxygenize(root = here::here())
}
\arguments{
\item{root}{The root of project.}
}
\value{
A logical vector of length 1.
}
\description{
This is the case if a new or replaced/removed line contains a roxygen2
comment in a file that is staged.
This function is only exported for use in hook scripts, but it's not intended
to be called by the end-user directly.
}
\examples{
\dontrun{
diff_requires_run_roxygenize()
}
}
\seealso{
Other hook script helpers: 
\code{\link{dirs_R.cache}()},
\code{\link{may_require_permanent_cache}()},
\code{\link{roxygen_assert_additional_dependencies}()},
\code{\link{roxygenize_with_cache}()}
}
\concept{hook script helpers}
