% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{enhance_mapping_special}
\alias{enhance_mapping_special}
\title{Enhance the mapping of text to the token "SPECIAL"}
\usage{
enhance_mapping_special(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
Map text corresponding to the token "SPECIAL" to a (more) unique token
description.
}
\keyword{internal}
