% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit.R
\name{enrich_terminals}
\alias{enrich_terminals}
\title{Enrich flattened parse table}
\usage{
enrich_terminals(flattened_pd, use_raw_indention = FALSE)
}
\arguments{
\item{flattened_pd}{A nested parse table that was turned into a flat parse
table using \code{\link[=extract_terminals]{extract_terminals()}}.}

\item{use_raw_indention}{Boolean indicating wheter or not the raw indention
should be used.}
}
\description{
Enriches a flattened parse table with terminals only. In particular, it is
possible to compute the line on which a token occurs and compute the exact
position a token will have on that line if it would be serialized.
}
