% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylerignore.R
\name{env_add_stylerignore}
\alias{env_add_stylerignore}
\title{Add positional information of token to next terminal}
\usage{
env_add_stylerignore(pd_flat)
}
\arguments{
\item{pd_flat}{A parse table.}
}
\description{
This is needed because at serialization time, we also have terminals only
and positional argument of non-terminals were already propagated to terminals
with \code{\link[=context_to_terminals]{context_to_terminals()}}.
}
\keyword{internal}
