% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\docType{data}
\name{env_current}
\alias{env_current}
\title{The elements that are added to this environment are:}
\format{An object of class \code{environment} of length 0.}
\usage{
env_current
}
\description{
The elements that are added to this environment are:
}
\details{
\itemize{
\item \code{parser_version}: Needed to dispatch between parser versions, see
\code{\link[=parser_version_set]{parser_version_set()}} for details.
\item \code{stylerignore}: A tibble with parse data containing tokens that fall within
a stylerignore sequence. This is used after serializing the flattened
parse table to apply the initial formatting to these tokens. See
\link{stylerignore} for details.
\item \code{any_stylerignore}: Whether there is any stylerignore marker. The idea is
to check early in the runtime if this is the case and then if so, take
as many short-cuts as possible. See \link{stylerignore} for details.
}
}
\keyword{internal}
