% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit.R
\name{extract_terminals}
\alias{extract_terminals}
\title{Extract terminal tokens}
\usage{
extract_terminals(pd_nested)
}
\arguments{
\item{pd_nested}{A nested parse table.}
}
\description{
Turns a nested parse table into a flat parse table. In particular it extracts
terminal tokens and the following attributes:
}
\details{
\itemize{
\item lag_newlines
\item indent
\item token
\item text
\item spaces
\item id
\item parent
\item line1
}
}
