% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-examples-find.R
\name{find_dont_seqs}
\alias{find_dont_seqs}
\title{Find \code{dontrun} and friend sequences}
\usage{
find_dont_seqs(bare)
}
\arguments{
\item{bare}{Bare code.}
}
\description{
Returns the indices of the lines that correspond to a \code{dontrun} or
friends sequence.
}
\keyword{internal}
