% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-create.R
\name{find_start_pos_id}
\alias{find_start_pos_id}
\title{Find legit starting value for a new positional id}
\usage{
find_start_pos_id(pd, pos, by, direction, after)
}
\arguments{
\item{pd}{A parse table.}

\item{pos}{The position where the new id should be inserted.}

\item{by}{By how much the reference \code{pos_id} should be increased / decreased
to create a new id.}

\item{after}{Boolean indicating whether it should be inserted after or before
\code{pos}.}
}
\description{
Looks at the current nest as well as into its children (if necessary) to make
sure the right id is returned. Otherise, ordering of tokens might not be
preserved.
}
