% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-replacement.R
\name{fix_quotes}
\alias{fix_quotes}
\title{Replace single quotes with double quotes}
\usage{
fix_quotes(pd_flat)
}
\arguments{
\item{pd_flat}{A flat parse table.}
}
\description{
Replace single quotes with double quotes
}
\details{
Depending on whether the backslash represents a character on its own or
whether it is used to create a special character, we either use:
\itemize{
\item \code{deparse(parse_text(...))}
\item \code{parse(paste("\"", ., "\""))}
}

To make sure esaping is done correctly.
}
\examples{
charToRaw("\\n") # one character
charToRaw("\\\\n") # two characters
charToRaw("\\\\\\n") # two characters
# put these into a file and style with style_file(),
# style_text() escapes backslash.
'here
is a string
'
'here\\nis a string
'
'here\\\\nis a string'
}
\keyword{internal}
