% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-tokens.R
\name{fix_quotes}
\alias{fix_quotes}
\title{Replace single quotes with double quotes}
\usage{
fix_quotes(pd_flat)
}
\arguments{
\item{pd_flat}{A flat parse table.}
}
\description{
We do not use \code{deparse()} as in previous implementations but \code{paste0()} since
the former approach escapes the reverse backslash in the line break character
\verb{\\\\n} whereas the solution with \code{paste0()} does not.
}
\examples{
style_text("'here
is a string
'")
}
\keyword{internal}
