% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-code.R
\name{get_knitr_pattern}
\alias{get_knitr_pattern}
\title{Get chunk pattern}
\usage{
get_knitr_pattern(lines, filetype = "Rmd")
}
\arguments{
\item{lines}{a character vector of lines from an Rmd file}
}
\description{
Determine a regex pattern for identifying R code chunks.
}
\keyword{internal}
