% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{get_parse_data}
\alias{get_parse_data}
\title{Obtain robust parse data}
\usage{
get_parse_data(text, include_text, ...)
}
\arguments{
\item{text}{The text to parse.}

\item{include_text}{Passed to \code{\link[utils:getParseData]{utils::getParseData()}} as \code{includeText}.}

\item{...}{Other arguments passed to \code{\link[utils:getParseData]{utils::getParseData()}}.}
}
\description{
Wrapper around \code{utils::getParseData(parse(text = text))} that returns a flat
parse table.
}
