% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{get_transformers}
\alias{get_transformers}
\title{get the transformer functions}
\usage{
get_transformers(strict = TRUE)
}
\arguments{
\item{strict}{A logical value indicating whether a set of strict
or not so strict transformer functions should be returned.}
}
\description{
get the transformer functions
}
