% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transformers.R
\name{get_transformers_nested}
\alias{get_transformers_nested}
\title{Get the transformer functions for nested styling}
\usage{
get_transformers_nested(scope = "tokens", strict = TRUE, indent_by = 2,
  start_comments_with_one_space = FALSE)
}
\arguments{
\item{scope}{The extent of manipulation. Can range from "none" (least
invasive) to "token" (most invasive). See 'Details'. This argument is a
vector of length one.}

\item{strict}{A logical value indicating whether a set of strict
or not so strict transformer functions should be returned.}

\item{indent_by}{How many spaces of indention should be inserted after
operators such as '('.}

\item{start_comments_with_one_space}{Whether or not comments should start
with only one space (see \code{\link[=start_comments_with_space]{start_comments_with_space()}}).}
}
\description{
Similar to \code{\link[=get_transformers_flat]{get_transformers_flat()}}, but additionally, returns some
functions needed due the fact that styling is done in a nested way.
}
\details{
The following options for \code{scope} are available.
\itemize{
\item "none": Performs no transformation at all.
\item "spaces": Manipulates spacing between token on the same line.
\item "line_breaks": In addition to "spaces", this option also manipulates
line breaks.
\item "tokens": In addition to "line_breaks", this option also manipulates
tokens.
}

As it becomes clear from this description, more invasive operations can only
be performed if all less invasive operations are performed too.
}
\seealso{
Other obtain transformers: \code{\link{get_transformers_flat}}
}
