% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{git_init}
\alias{git_init}
\title{Initiate git and configure it}
\usage{
git_init(path = ".")
}
\description{
In particular, to avoid CRAN errors
\href{https://github.com/lorenzwalthert/precommit/issues/320}{lorenzwalthert/precommit#320}.
}
\keyword{internal}
