% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-alignment.R
\name{has_correct_spacing_around_eq_sub}
\alias{has_correct_spacing_around_eq_sub}
\title{At least one space around \code{EQ_SUB}}
\usage{
has_correct_spacing_around_eq_sub(pd_sub)
}
\description{
At least one space around \code{EQ_SUB}
}
\keyword{internal}
