% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_rmd.R
\name{identify_chunks}
\alias{identify_chunks}
\title{Identify chunks within Rmd contents}
\usage{
identify_chunks(lines)
}
\arguments{
\item{lines}{a character vector of lines from an Rmd file}
}
\description{
Identifies the code and text chunks within an Rmd file, and returns these
as a nested list.
}
