% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{identify_insufficiently_parsed_stings}
\alias{identify_insufficiently_parsed_stings}
\title{Indentify strings that were not fully parsed}
\usage{
identify_insufficiently_parsed_stings(pd, text)
}
\arguments{
\item{pd}{A parse table.}

\item{text}{The initial code to style.}
}
\description{
Indentifies strings that were not fully parsed due to their vast length.
}
\details{
The meaning of the variable \code{is_problematic_string} in the source code
changes from "all strings" to "all problematic strings", is partly
missleading and this approach was choosen for performance reasons only.
}
\keyword{internal}
