% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-code.R
\name{identify_raw_chunks}
\alias{identify_raw_chunks}
\title{Identifies raw Rmd or Rnw code chunks}
\usage{
identify_raw_chunks(lines, filetype,
  engine_pattern = get_engine_pattern())
}
\arguments{
\item{lines}{A character vector of lines from an Rmd or Rnw file.}

\item{filetype}{A string indicating the filetype - either 'Rmd' or 'Rnw'.}

\item{engine_pattern}{A regular expression that must match the engine name.}
}
\description{
Raw in the sense that these chunks don't contain pure R code, but they
contain a header and footer of markdown. Only code chunks that have an engine
whose name matches \code{engine-pattern} are considered as R code.
}
\keyword{internal}
