% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_precommit}
\alias{install_precommit}
\title{Install pre-commit on your system.}
\usage{
install_precommit()
}
\description{
<<<<<<< HEAD
This installs pre-commit in the conda environment r-precommit. It
=======
This installs pre-commit in the conda environment r-precommit It
>>>>>>> switch to env r-precommit to avoid conflicts
will be available to use across different git repositories.
}
