% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_precommit}
\alias{install_precommit}
\title{Install pre-commit on your system.}
\usage{
install_precommit()
}
\description{
This installs pre-commit in the conda environment r-precommit. It
will be available to use across different git repositories.
}
