% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_precommit}
\alias{install_precommit}
\title{Install pre-commit on your system.}
\usage{
install_precommit(force = FALSE)
}
\arguments{
\item{force}{Whether or not to force a re-installation.}
}
\value{
The path to the pre-commit executable (invisibly).
}
\description{
This installs pre-commit in the conda environment r-precommit. It
will be available to use across different git repositories.
}
\examples{
\dontrun{
install_precommit()
}
}
