% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{is_cached}
\alias{is_cached}
\title{Check if text is cached}
\usage{
is_cached(text, transformers, cache_dir = cache_dir_default())
}
\arguments{
\item{text, transformers}{Passed to \code{\link[=cache_make_key]{cache_make_key()}} to generate a key.}

\item{cache_dir}{The caching directory relative to the \code{.Rcache} root to
look for a cached value.}
}
\description{
This boils down to check if the hash exists at the caching dir as a file.
}
\keyword{internal}
