% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-indention.R
\name{is_single_indent_function_declaration}
\alias{is_single_indent_function_declaration}
\title{Is the function declaration authored with standard single/double indentation?}
\usage{
is_single_indent_function_declaration(pd, indent_by = 2L)
}
\arguments{
\item{pd}{A parse table.}

\item{indent_by}{How many spaces of indention should be inserted after
operators such as '('.}
}
\description{
Assumes you already checked if it's a function declaration with
\code{is_function_declaration}. It returns \code{TRUE} if the first formal argument
starts on a new line AND either the closing parenthesis \verb{)} also starts on a
new line or the first argument is indented by standard single/double
indentation (\verb{<= 2 * indent_by} spaces). This robustly identifies standard
multi-line headers while correctly returning \code{FALSE} for aligned/hanging
indents (where arguments start on the first line or share the line with \verb{)}).
}
\keyword{internal}
