% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-indention.R
\name{is_single_indent_function_declaration}
\alias{is_single_indent_function_declaration}
\title{Is the function declaration single indented?}
\usage{
is_single_indent_function_declaration(pd, indent_by = 2L)
}
\arguments{
\item{pd}{A parse table.}

\item{indent_by}{How many spaces of indention should be inserted after
operators such as '('.}
}
\description{
Assumes you already checked if it's a function with
\code{is_function_declaration}. "single indented" means the formals are on the
line after \verb{function(} and indented one further level, and the closing \verb{)}
is also on its own line. The alternative compliant style within the style
guide is "hanging indented" where formals share the line with \verb{function(},
any subsequent lines of formals are indented relative to that \code{(}, and
the closing \verb{)} shares a line with the last formal argument.
}
\keyword{internal}
