% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{local_test_setup}
\alias{local_test_setup}
\title{Establish testing setup for current environment}
\usage{
local_test_setup(testthat_cache = FALSE, .local_envir = parent.frame())
}
\description{
Establish testing setup for current environment
}
\details{
\itemize{
\item make styler quiet.
}
}
