% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{local_test_setup}
\alias{local_test_setup}
\title{Testing utilities}
\usage{
local_test_setup(
  git = TRUE,
  use_precommit = FALSE,
  package = FALSE,
  quiet = TRUE,
  autoupdate = FALSE,
  ...,
  .local_envir = parent.frame()
)
}
\arguments{
\item{git}{Whether or not to init git in the local directory.}

\item{autoupdate}{Whether or not to run \code{\link[=autoupdate]{autoupdate()}} as part of this
fixture.}

\item{.local_envir}{\verb{[environment]}\cr The environment to use for scoping.}

\item{use_precommmit}{Whether or not to \code{\link[=use_precommit]{use_precommit()}}.}
}
\description{
Similar to the \code{local_()} family from \code{{withr}}, this function creates a
temporary directory and optionally initiates git and pre-commit in it.
}
\keyword{internal}
