% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-define.R
\name{lookup_new_special}
\alias{lookup_new_special}
\title{Lookup which new tokens were created from "SPECIAL"}
\usage{
lookup_new_special(regex = NA)
}
\arguments{
\item{regex}{A regular expression pattern to search for.}
}
\description{
Lookup which new tokens were created from "SPECIAL"
}
\keyword{internal}
