% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{make_transformer}
\alias{make_transformer}
\title{Closure to return a transformer function}
\usage{
make_transformer(transformers, flat = FALSE)
}
\arguments{
\item{transformers}{A list of transformer functions that operate on parse
tables.}

\item{flat}{Whether to do the styling with a flat approach or with a nested
approach.
TODO verify_roundtrip}
}
\description{
This function takes a list of transformer functions as input and
returns a function that can be applied to character strings
that should be transformed.
}
