% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{make_transformer}
\alias{make_transformer}
\title{TODO:
\itemize{
\item Walk tree defined by \code{leaves}, compute whitespace
\item Perform all transformations on hierarchical structure
\itemize{
\item Compute text for a sub-element
}
\item Compute indentation
\itemize{
\item Braces
\item Function calls
\item Function definitions
}
\item Remove \code{includeText = TRUE}
}}
\usage{
make_transformer(transformers)
}
\description{
TODO:
\itemize{
\item Walk tree defined by \code{leaves}, compute whitespace
\item Perform all transformations on hierarchical structure
\itemize{
\item Compute text for a sub-element
}
\item Compute indentation
\itemize{
\item Braces
\item Function calls
\item Function definitions
}
\item Remove \code{includeText = TRUE}
}
}
