% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{may_require_permanent_cache}
\alias{may_require_permanent_cache}
\title{Issue a warning if \code{{R.cache}} uses temporary cache only}
\usage{
may_require_permanent_cache(temp_cache_is_enough = FALSE)
}
\arguments{
\item{temp_cache_is_enough}{Whether a temporary cache is accepted or not. \code{TRUE}
means no warning will be issued, \code{FALSE} means a warning will be issued if
no permanent cache is available.}
}
\description{
This function is only exported for use in hook scripts, but it's not intended
to be called by the end-user directly.
}
\seealso{
Other hook script helpers: 
\code{\link{diff_requires_run_roxygenize}()},
\code{\link{dirs_R.cache}()},
\code{\link{roxygen_assert_additional_dependencies}()},
\code{\link{roxygenize_with_cache}()}
}
\concept{hook script helpers}
