% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{may_require_permanent_cache}
\alias{may_require_permanent_cache}
\title{Issue a warning if \code{{R.cache}} uses temporary cache only}
\usage{
may_require_permanent_cache(temp_cache_is_enough = FALSE)
}
\arguments{
\item{temp_cache_is_enough}{ignored.}
}
\description{
This function used to check if a permanent cache was available and issue a
warning if not, but since {R.cache} version \verb{0.15.0} (release date
2021-04-27), a permanent directory will be used automatically, so this check
if redundant. the function is kept in the package for compatibility, i.e.
if someone updates the R package {precommit} but not the hook revisions.
}
\seealso{
Other hook script helpers: 
\code{\link{diff_requires_run_roxygenize}()},
\code{\link{dirs_R.cache}()},
\code{\link{roxygen_assert_additional_dependencies}()},
\code{\link{roxygenize_with_cache}()}
}
\concept{hook script helpers}
