% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_pd.R
\name{needs_indention}
\alias{needs_indention}
\title{Check whether indention is needed}
\usage{
needs_indention(pd, token = "'('")
}
\arguments{
\item{pd}{A parse table.}

\item{token}{Which token the check should be based on.}
}
\value{
returns \code{TRUE} if indention is needed, \code{FALSE} otherwise. Indention
is needed:
* if \code{token} occurs in \code{pd}.
* if there is no child that starts on the same line as \code{token} and
indents.

\code{TRUE} if indention is needed, \code{FALSE} otherwise.
}
\description{
Check whether indention is needed
}
