% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_pd.R
\name{needs_indention}
\alias{needs_indention}
\title{Check whether indention is needed}
\usage{
needs_indention(pd, opening)
}
\arguments{
\item{pd}{A parse table.}

\item{opening}{the index of the opening parse table. Since always computed
before this function is called, it is included as an argument so it does
not have to be recomputed.}
}
\value{
returns \code{TRUE} if indention is needed, \code{FALSE} otherwise. Indention
is needed if and only if:
* the opening token is not \code{NA}.
* if there is a multi-line token before the first line break.

\code{TRUE} if indention is needed, \code{FALSE} otherwise.
}
\description{
Check whether indention is needed
}
