% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent.R
\name{needs_indention_one}
\alias{needs_indention_one}
\title{Check whether indention is needed}
\usage{
needs_indention_one(pd, potential_trigger)
}
\arguments{
\item{pd}{A parse table.}

\item{potential_trigger}{the index of the token in the parse table
for which it should be checked whether it should trigger indention.}
}
\value{
Returns \code{TRUE} if indention is needed, \code{FALSE} otherwise.

\code{TRUE} if indention is needed, \code{FALSE} otherwise.
}
\description{
Indention is needed if and only if there is no multi-line token between the
trigger and the first line break.
}
