% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_config}
\alias{open_config}
\alias{open_wordlist}
\title{Open pre-commit related files}
\usage{
open_config(path_root = here::here())

open_wordlist(path_root = here::here())
}
\arguments{
\item{path_root}{The path to the root directory of your project.}
}
\description{
Open pre-commit related files
}
\details{
\itemize{
\item \code{open_config()}: opens the pre-commit config file.
\item \code{open_wordlist()}: opens the the WORDLIST file for the check-spelling hook
in inst/WORDLIST.
}
}
\seealso{
Other helpers: 
\code{\link{use_precommit}()}
}
\concept{helpers}
