% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_config}
\alias{open_config}
\alias{open_wordlist}
\title{Open pre-commit related files}
\usage{
open_config(root = here::here())

open_wordlist(root = here::here())
}
\arguments{
\item{root}{The path to the root directory of your project.}
}
\value{
\code{NULL} (invisibly). The function is called for its side effects.
}
\description{
Open pre-commit related files
}
\details{
\itemize{
\item \code{open_config()}: opens the pre-commit config file.
\item \code{open_wordlist()}: opens the the WORDLIST file for the check-spelling hook
in inst/WORDLIST.
}
}
\examples{
\dontrun{
open_config()
}
\dontrun{
open_wordlist()
}
}
\seealso{
Other helpers: 
\code{\link{use_precommit}()}
}
\concept{helpers}
