% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-block.R
\name{parse_transform_serialize_r_block}
\alias{parse_transform_serialize_r_block}
\title{Parse, transform and serialize a nested parse table}
\usage{
parse_transform_serialize_r_block(pd_nested, start_line, transformers)
}
\arguments{
\item{pd_nested}{A block of the nested parse table.}

\item{start_line}{The line number on which the code starts.}

\item{transformers}{A list of \emph{named} transformer functions}
}
\description{
We process blocks of nested parse tables for speed. See \code{\link[=cache_find_block]{cache_find_block()}}
for details on how a top level nest is split into blocks.
}
\keyword{internal}
