% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{path_derive_precommit_exec}
\alias{path_derive_precommit_exec}
\title{Derive the path to the pre-commit executable}
\usage{
path_derive_precommit_exec()
}
\description{
Returns "" if search was not successful, the path otherwise.
}
\section{Heuristic}{

\itemize{
\item First check if there is an executable on the \verb{$PATH} using
\code{\link[=path_derive_precommit_exec_path]{path_derive_precommit_exec_path()}}
\item Search os dependent for other possible locations for common installation
methods.
\item If not, check if we can find one in a conda environment with
\code{\link[=path_derive_precommit_exec_conda]{path_derive_precommit_exec_conda()}}. Do this last as it's the slowest.
}
}

\keyword{internal}
