% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{path_precommit_exec}
\alias{path_precommit_exec}
\alias{path_pre_commit_exec}
\title{Locate the pre-commit executable}
\usage{
path_precommit_exec(check_if_exists = TRUE)

path_pre_commit_exec(check_if_exists = TRUE)
}
\arguments{
\item{check_if_exists}{Whether or not to make sure the returned path also
exists.}
}
\value{
A character vector of length one with the path to the pre-commit executable.
}
\description{
\code{\link[=path_precommit_exec]{path_precommit_exec()}} simply reads the R option \code{precommit.executable}, ]
\code{\link[=path_derive_precommit_exec]{path_derive_precommit_exec()}} uses a heuristic to derive it from scratch.
}
\examples{
\dontrun{
path_precommit_exec()
}
\dontrun{
path_pre_commit_exec()
}
}
