% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_pd.R
\name{pd_has_token}
\alias{pd_has_token}
\title{Check whether a parse table contains a token}
\usage{
pd_has_token(pd, token)
}
\arguments{
\item{pd}{A parse table.}

\item{token}{The token for which it should be checked whether it is in the
parse table.}
}
\value{
\code{TRUE} if the token is in the parse table, \code{FALSE} otherwise.
}
\description{
Check whether a parse table contains a token
}
