% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-examples-parse.R
\name{post_parse_roxygen}
\alias{post_parse_roxygen}
\title{Changing the line definition}
\usage{
post_parse_roxygen(raw)
}
\arguments{
\item{raw}{Raw code to post-process.}
}
\description{
Input: New line denoted with \verb{\\\\n}. Lines can span across elements.
Output: Each element in the vector is one line.
}
\keyword{internal}
