% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precommit-package.R
\docType{package}
\name{precommit-package}
\alias{precommit}
\alias{precommit-package}
\title{precommit: Pre-Commit Hooks}
\description{
Useful git hooks for R building on top of the
    multi-language framework 'pre-commit' for hook management. This
    package provides git hooks for common tasks like formatting files with
    styler or spell checking as well as wrapper functions to access the
    'pre-commit' executable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lorenzwalthert.github.io/precommit/}
  \item \url{https://github.com/lorenzwalthert/precommit}
}

}
\keyword{internal}
