% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{prettify_any}
\alias{prettify_any}
\title{Prettify R code in current working directory}
\usage{
prettify_any(transformers, recursive)
}
\arguments{
\item{transformers}{A list of transformer functions to be applied to the
files in \code{path}.}

\item{recursive}{A logical value indicating whether or not files in subdirectories
should be styled as well.}
}
\description{
This is a helper function for style_src.
}
\keyword{internal}
