% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{prettify_one}
\alias{prettify_one}
\title{Prettify one R file}
\usage{
prettify_one(transformers, flat, path)
}
\arguments{
\item{transformers}{A list with functions to be applied to the parsed data.}

\item{flat}{Whether to do the styling with a flat approach or with a nested
approach.}

\item{path}{The path to a file that should be styled.}
}
\description{
This is a helper function for style_src.
}
