% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{prettify_one}
\alias{prettify_one}
\title{Prettify one R file}
\usage{
prettify_one(transformers, path)
}
\arguments{
\item{transformers}{A set of transformer functions.}

\item{path}{The path to a file that should be styled.}
}
\description{
This is a helper function for style_dir.
}
