% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{regex_strcode}
\alias{regex_strcode}
\title{Return the regular expressions from the strcode package}
\usage{
regex_strcode()
}
\description{
Returns a character vector with the patterns of code separator and title
for the three levels of granularity.
}
