% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-replacement.R
\name{remove_quotes_in_library_call}
\alias{remove_quotes_in_library_call}
\title{Removes quotes in the argument \code{package} in library()}
\usage{
remove_quotes_in_library_call(pd)
}
\description{
Function finds the position in the parse table that corresponds to the
argument \code{package}, if any. The argument corresponds to the named argument
\code{package} or the first unnamed argument otherwise.
Note that for fiding the argument \code{package}, we can't use \code{match.call()}
because it does not
preserve the order. Then, the quotes are removed if and only if the
package argument is a string and not a symbol plus \code{character.only} is set
to \code{FALSE}. To determine the value of \code{character.only}, we can use
\code{\link[=match.call]{match.call()}}.
}
\examples{
style_text("library('tidyverse', character.only = TRUE)")
style_text("library(tidyverse, character.only = TRUE)")
}
\keyword{internal}
