% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{roxygen2md}
\alias{roxygen2md}
\title{Convert from Rd to Markdown in roxygen2 comments}
\usage{
roxygen2md(pkg = ".")
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package}
}
\description{
Performs various substitutions in all \code{.R} files in a package.
Carefully examine the results after running this function!
}

