% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{run_test_impl}
\alias{run_test_impl}
\title{Implement a test run}
\usage{
run_test_impl(path_executable, path_candidate, error_msg, cmd_args, copy)
}
\arguments{
\item{path_executable}{The path to the executable bash script.}

\item{path_candidate}{The path to a file that should be modified by the
executable.}

\item{copy}{Path with files to copy to the temp directory where the test
is run.}
}
\description{
Implement a test run
}
