% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{run_test_impl}
\alias{run_test_impl}
\title{Implement a test run}
\usage{
run_test_impl(
  path_executable,
  path_candidate,
  error_msg,
  msg,
  cmd_args,
  artifacts,
  file_transformer,
  env
)
}
\arguments{
\item{path_executable}{The path to the executable bash script.}

\item{path_candidate}{The path to a file that should be modified by the
executable.}

\item{error_msg}{An expected error message. If no error is expected, this
can be \code{NULL}. In that case, the \code{comparator} is applied.}

\item{msg}{The expected stdout message. If \code{NULL}, this check is omitted.}

\item{cmd_args}{More arguments passed to the file. Pre-commit handles it as
described \href{https://pre-commit.com/#arguments-pattern-in-hooks}{here}.}

\item{artifacts}{Path with artifact files to copy to the temp directory root where
the test is run. If you don't target the root, this can be a named vector
of length one where the name is the target location relative to the
temporary location and the value is the source of the file.}

\item{env}{The environment variables to set with \code{\link[base:system2]{base::system2()}}.}
}
\description{
Implement a test run
}
\keyword{internal}
