% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{run_test_impl}
\alias{run_test_impl}
\title{Implement a test run}
\usage{
run_test_impl(path_executable, path_candidate, error_msg, cmd_args, copy)
}
\arguments{
\item{path_executable}{The path to the executable bash script.}

\item{path_candidate}{The path to a file that should be modified by the
executable.}

\item{error_msg}{An expected error message. If no error is expected, this
can be \code{NULL}. In that case, the \code{comparator} is applied.}

\item{cmd_args}{More arguments passed to the file. Pre-commit handles it as
described \href{https://pre-commit.com/#arguments-pattern-in-hooks}{here}.}

\item{copy}{Path with files to copy to the temp directory where the test
is run. If the target destination relative to the temp dir where the hook
is tested is not identical to the path from where a file should be copied,
you can pass a named vector. The name is the target directory relative to
the temp directory where the hook is executed (the temp directory will be
the working directory at that time) and the value is the path that points
to the place where the artifact is currently stored.}
}
\description{
Implement a test run
}
\keyword{internal}
