% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.R
\name{serialize_parse_data_nested_helper}
\alias{serialize_parse_data_nested_helper}
\title{Serialize a nested parse table}
\usage{
serialize_parse_data_nested_helper(pd_nested, pass_indent)
}
\arguments{
\item{pd_nested}{A nested parse table.}

\item{pass_indent}{Level of indention of a token.}
}
\value{
A character vector with all terminal tokens in \code{pd_nested} plus
the appropriate amount of white spaces and line breaks are inserted between
them.
}
\description{
Helper function that recursively extracts terminals from a nested tibble.
}
