% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-line_break.R
\name{set_line_break_if_call_is_multi_line}
\alias{set_line_break_if_call_is_multi_line}
\alias{set_line_break_after_opening_if_call_is_multi_line}
\alias{set_line_break_before_closing_if_call_is_multi_line}
\alias{remove_line_break_in_empty_fun_call}
\title{Set line break for multi-line function calls}
\usage{
set_line_break_after_opening_if_call_is_multi_line(pd, except_token = NULL)

set_line_break_before_closing_if_call_is_multi_line(pd)

remove_line_break_in_empty_fun_call(pd)
}
\arguments{
\item{pd}{A parse table.}

\item{except}{A character vector with tokens before "'('" that do not}
}
\description{
Set line break for multi-line function calls
}
\section{Functions}{
\itemize{
\item \code{set_line_break_after_opening_if_call_is_multi_line}: Sets line break after
opening parenthesis.

\item \code{set_line_break_before_closing_if_call_is_multi_line}: Sets line break before
closing parenthesis.
}}

