% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-spaces.R
\name{set_space_between_eq_sub_and_comma}
\alias{set_space_between_eq_sub_and_comma}
\title{Set space between EQ_SUB and "','"}
\usage{
set_space_between_eq_sub_and_comma(pd)
}
\arguments{
\item{pd}{A parse table.}
}
\description{
Set space between EQ_SUB and "','"
}
\keyword{internal}
