% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{set_spaces}
\alias{set_spaces}
\title{Helper for setting spaces}
\usage{
set_spaces(spaces_after_prefix, force_one)
}
\arguments{
\item{spaces_after_prefix}{An integer vector with the number of spaces
after the prefix.}

\item{force_one}{Whether spaces_after_prefix should be set to one in all
cases.}
}
\value{
An integer vector of length spaces_after_prefix, which is either
one (if \code{force_one = TRUE}) or \code{space_after_prefix} with all values
below one set to one.

Numeric vector indicating the number of spaces.
}
\description{
Helper for setting spaces
}
\keyword{internal}
