% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent-token-dependent.R
\name{set_token_dependent_indention_one}
\alias{set_token_dependent_indention_one}
\title{Set the token-dependent indention}
\usage{
set_token_dependent_indention_one(flattened_pd, target_index)
}
\arguments{
\item{flattened_pd}{A flattened parse table.}

\item{target_index}{The index of the token in \code{flattened_pd} that has the
indention we want to apply to other tokens.}
}
\value{
A parse table with updated indention for \code{target_index}.
}
\description{
Set the token-dependent indention for one target index.
}
