% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-spaces.R
\name{start_comments_with_space}
\alias{start_comments_with_space}
\title{Start comments with a space}
\usage{
start_comments_with_space(pd, force_one = FALSE)
}
\arguments{
\item{pd}{A parse table.}

\item{force_one}{Whether or not to force one space or allow multiple spaces.}
}
\description{
Forces comments to start with a space, that is, after the regular expression
\verb{#+['\\\\*]}, at least one space must follow if the comment is \emph{non-empty}, i.e
there is not just spaces within the comment. Multiple spaces may be legit for
indention in some situations.
}
\section{Exceptions}{

Spaces won't be added to comments when they are:
\itemize{
\item shebangs
\item code chunk headers
\item xaringan markers
}
}

\keyword{internal}
