% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{style_dir}
\alias{style_dir}
\title{Prettify arbitrary R code}
\usage{
style_dir(path = ".", ..., style = tidyverse_style,
  transformers = style(...), recursive = TRUE)
}
\arguments{
\item{path}{Path to a directory with files to transform.}

\item{...}{Passed on to the \code{style} function.}

\item{style}{A function that creates a style guide to use, by default
\code{\link[=tidyverse_style]{tidyverse_style()}} (without the parentheses). Will not be used further
except to construct the argument \code{transformers}.}

\item{transformers}{A set of transformer functions.}

\item{recursive}{A logical value indicating whether or not files in subdirectories
of \code{path} should be styled as well.}
}
\description{
Performs various substitutions in all \code{.R} files in a directory.
Carefully examine the results after running this function!
}
\seealso{
Other stylers: \code{\link{style_file}},
  \code{\link{style_pkg}}, \code{\link{style_text}}
}
