% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{style_dir}
\alias{style_dir}
\title{Prettify arbitrary R code}
\usage{
style_dir(path = ".", flat = FALSE, recursive = TRUE,
  transformers = get_transformers(flat = flat))
}
\arguments{
\item{path}{Path to a directory with files to transform.}

\item{flat}{Whether to do the styling with a flat approach or with a nested
approach.}

\item{recursive}{A logical value indicating whether or not files in subdirectories
of \code{path} should be styled as well.}

\item{transformers}{A list with functions to be applied to the parsed data.}
}
\description{
Performs various substitutions in all \code{.R} files in a directory.
Carefully examine the results after running this function!
}
\seealso{
Other stylers: \code{\link{style_file}},
  \code{\link{style_pkg}}, \code{\link{style_text}}
}
