% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{style_dir}
\alias{style_dir}
\title{Prettify arbitrary R code}
\usage{
style_dir(path = ".", ..., style = tidyverse_style,
  transformers = style(...), recursive = TRUE, exclude_files = NULL)
}
\arguments{
\item{path}{Path to a directory with files to transform.}

\item{...}{Arguments passed on to the \code{style} function.}

\item{style}{A function that creates a style guide to use, by default
\code{\link[=tidyverse_style]{tidyverse_style()}} (without the parentheses). Not used
further except to construct the argument \code{transformers}. See
\code{\link[=style_guides]{style_guides()}} for details.}

\item{transformers}{A set of transformer functions. This argument is most
conveniently constructed via the \code{style} argument and \code{...}. See
'Examples'.}

\item{recursive}{A logical value indicating whether or not files in subdirectories
of \code{path} should be styled as well.}

\item{exclude_files}{Character vector with paths to files that should be
excluded from styling.}
}
\description{
Performs various substitutions in all \code{.R} files in a directory.
Carefully examine the results after running this function!
}
\section{Warning}{

This function overwrites files (if styling results in a change of the
code to be formatted). It is strongly suggested to only style files
that are under version control or to create a backup copy.
}

\seealso{
Other stylers: \code{\link{style_files}},
  \code{\link{style_file}}, \code{\link{style_pkg}},
  \code{\link{style_text}}, \code{\link{styler_addins}}
}
