% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{style_dir}
\alias{style_dir}
\title{Prettify arbitrary R code}
\usage{
style_dir(path = ".", ..., style = tidyverse_style,
  transformers = style(...), recursive = TRUE)
}
\arguments{
\item{path}{Path to a directory with files to transform.}

\item{...}{Passed on to the \code{style} function.}

\item{style}{The unquoted name of a style guide to use. Will not be used
further except to construct the argument \code{transformers}. See
\code{\link[=style_guides]{style_guides()}} for details.}

\item{transformers}{A set of transformer functions.}

\item{recursive}{A logical value indicating whether or not files in subdirectories
of \code{path} should be styled as well.}
}
\description{
Performs various substitutions in all \code{.R} files in a directory.
Carefully examine the results after running this function!
}
\seealso{
Other stylers: \code{\link{style_file}},
  \code{\link{style_pkg}}, \code{\link{style_text}}
}
