% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{style_file}
\alias{style_file}
\title{Style a file}
\usage{
style_file(path, flat = FALSE, transformers = get_transformers(flat = flat))
}
\arguments{
\item{path}{A path to a file to style.}

\item{flat}{Whether to do the styling with a flat approach or with a nested
approach.}

\item{transformers}{A list with functions to be applied to the parsed data.}
}
\description{
Performs various substitutions in the \code{.R} file specified.
Carefully examine the results after running this function!
}
\seealso{
Other stylers: \code{\link{style_pkg}},
  \code{\link{style_src}}, \code{\link{style_text}}
}
