% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{style_file}
\alias{style_file}
\title{Style a file}
\usage{
style_file(path, ..., style = tidyverse_style, transformers = style(...))
}
\arguments{
\item{path}{A path to a file to style.}

\item{...}{Passed on to the \code{style} function.}

\item{style}{The unquoted name of a style guide to use. Will not be used
further except to construct the argument \code{transformers}. See
\code{\link[=style_guides]{style_guides()}} for details.}

\item{transformers}{A set of transformer functions.}
}
\description{
Performs various substitutions in the \code{.R} file specified.
Carefully examine the results after running this function!
}
\seealso{
Other stylers: \code{\link{style_dir}},
  \code{\link{style_pkg}}, \code{\link{style_text}}
}
