% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{style_pkg}
\alias{style_pkg}
\title{Prettify R source code}
\usage{
style_pkg(pkg = ".", ..., style = tidyverse_style,
  transformers = style(...))
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package}

\item{...}{Passed on to \code{\link[=get_transformers]{get_transformers()}}}

\item{transformers}{A list with functions to be applied to the parsed data.}
}
\description{
Performs various substitutions in all \code{.R} files in a package
(code and tests).
Carefully examine the results after running this function!
}
\seealso{
Other stylers: \code{\link{style_dir}},
  \code{\link{style_file}}, \code{\link{style_text}}
}
