% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{style_pkg}
\alias{style_pkg}
\title{Prettify R source code}
\usage{
style_pkg(pkg = ".", transformers = get_transformers())
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package}
}
\description{
Performs various substitutions in all \code{.R} files in a package
(code and tests).
Carefully examine the results after running this function!
}
