% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{style_pkg}
\alias{style_pkg}
\title{Prettify R source code}
\usage{
style_pkg(pkg = ".", ..., style = tidyverse_style,
  transformers = style(...))
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package.}

\item{...}{Passed on to the \code{style} function.}

\item{style}{A function that creates a style guide to use, by default
\code{\link[=tidyverse_style]{tidyverse_style()}} (without the parentheses). Will not be used further
except to construct the argument \code{transformers}.}

\item{transformers}{A set of transformer functions.}
}
\description{
Performs various substitutions in all \code{.R} files in a package
(code and tests).
Carefully examine the results after running this function!
}
\seealso{
Other stylers: \code{\link{style_dir}},
  \code{\link{style_file}}, \code{\link{style_text}}
}
