% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{style_pkg}
\alias{style_pkg}
\title{Prettify R source code}
\usage{
style_pkg(pkg = ".", ..., style = tidyverse_style,
  transformers = style(...))
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package.}

\item{...}{Passed on to the \code{style} function.}

\item{style}{The unquoted name of a style guide to use. Will not be used
further except to construct the argument \code{transformers}. See
\code{\link[=style_guides]{style_guides()}} for details.}

\item{transformers}{A set of transformer functions.}
}
\description{
Performs various substitutions in all \code{.R} files in a package
(code and tests).
Carefully examine the results after running this function!
}
\examples{
\dontrun{
style_pkg()
}
}
\seealso{
Other stylers: \code{\link{style_dir}},
  \code{\link{style_file}}, \code{\link{style_text}}

Other stylers: \code{\link{style_dir}},
  \code{\link{style_file}}, \code{\link{style_text}}
}
