% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-roxygen-examples.R
\name{style_roxygen_dontrun_code_examples_one}
\alias{style_roxygen_dontrun_code_examples_one}
\title{Given a code segment is dontrun or run, style it}
\usage{
style_roxygen_dontrun_code_examples_one(code_segment, transformers, is_dontrun)
}
\arguments{
\item{code_segment}{A character vector with code to style.}

\item{transformers}{A list of \emph{named} transformer functions}

\item{is_dontrun}{Whether the segment to process is a dontrun segemnt or not.}
}
\description{
Given a code segment is dontrun or run, style it
}
\keyword{internal}
