% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{style_selection}
\alias{style_selection}
\title{Styles the highlighted selection in a \code{.R} or \code{.Rmd} file.}
\usage{
style_selection()
}
\description{
Styles the highlighted selection in a \code{.R} or \code{.Rmd} file.
}
\keyword{internal}
