% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{style_text}
\alias{style_text}
\title{Style a string}
\usage{
style_text(text, flat = FALSE, transformers = if (flat)
  get_transformers_flat() else get_transformers_nested())
}
\arguments{
\item{text}{A character vector with text to style.}

\item{transformers}{A list with functions to be applied to the parsed data.}
}
\description{
Styles a character vector
}
