% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws.R
\name{style_text}
\alias{style_text}
\title{Style a string}
\usage{
style_text(text, ..., style = tidyverse_style, transformers = style(...))
}
\arguments{
\item{text}{A character vector with text to style.}

\item{...}{Passed on to the \code{style} function.}

\item{style}{The unquoted name of a style guide to use. Will not be used
further except to construct the argument \code{transformers}. See
\code{\link[=style_guides]{style_guides()}} for details.}

\item{transformers}{A set of transformer functions.}
}
\description{
Styles a character vector. Each element of the character vector corresponds
to one line of code.
}
\examples{
style_text("call( 1)")
style_text("1    + 1", strict = FALSE)
}
\seealso{
Other stylers: \code{\link{style_dir}},
  \code{\link{style_file}}, \code{\link{style_pkg}}
}
