% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_transformers_drop}
\alias{test_transformers_drop}
\title{Test \code{transformers_drop} for consistency}
\usage{
test_transformers_drop(transformers)
}
\arguments{
\item{transformers}{The output of \code{\link[=create_style_guide]{create_style_guide()}} we want to test.}
}
\description{
Check if the argument \code{transformers_drop} in \code{\link[=create_style_guide]{create_style_guide()}} is
consistent with the transformers specified in that function.
}
\keyword{internal}
