% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{text_to_flat_pd}
\alias{text_to_flat_pd}
\title{Creates a flat parse table with minimal initialization}
\usage{
text_to_flat_pd(text, transformers, more_specs)
}
\arguments{
\item{text}{The text to parse.}

\item{transformers}{Passed to \code{\link[=cache_make_key]{cache_make_key()}} to generate a key.}

\item{more_specs}{Passed to \code{\link[=cache_make_key]{cache_make_key()}} to generate a key.}
}
\description{
Creates a flat parse table with minimal initialization and makes the parse
table shallow where appropriate.
}
\details{
This includes:
\itemize{
\item token before and after.
\item stylerignore attribute.
\item caching attributes.
}

Note that the parse table might be shallow if caching is enabled and some
values are cached.
}
\keyword{internal}
