% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{tokenize}
\alias{tokenize}
\title{Obtain token table from text}
\usage{
tokenize(text)
}
\arguments{
\item{text}{The text to parse.}
}
\value{
A flat parse table
}
\description{
\code{\link[utils:getParseData]{utils::getParseData()}} is used to obtain a flat parse table from \code{text}.
}
\details{
Apart from the columns provided by \code{utils::getParseData()}, the following
columns are added:
\itemize{
\item A column "short" with the first five characters of "text".
\item A column "pos_id" for (positional id) which can be used for sorting
(because "id" cannot be used in general). Note that the nth value of this
column corresponds to n as long as no tokens are inserted.
\item A column "child" that contains \emph{nest}s.
}
}
\keyword{internal}
