% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialized_tests.R
\name{transform_and_check}
\alias{transform_and_check}
\title{Transform a file an check the result}
\usage{
transform_and_check(in_item, out_item, in_name = in_item,
  out_name = out_item, transformer, write_back)
}
\arguments{
\item{in_item}{An path to an file to transform.}

\item{out_item}{The path to a file that contains the expected result.}

\item{in_name}{The label of the in_item, defaults to \code{in_item}.}

\item{out_name}{The label of the out_item, defaults to \code{out_item}.}

\item{transformer}{A function to apply to the content of \code{in_item}.}

\item{write_back}{Whether the results of the transformation should be written
to the output file.}
}
\description{
Transform an file and check whether it is identical to a reference.
}
