% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_rmd.R
\name{transform_code}
\alias{transform_code}
\title{Transform code from R or Rmd files}
\usage{
transform_code(path, fun, verbose = FALSE, ...)
}
\arguments{
\item{path}{A vector of file paths.}

\item{fun}{A function that returns a character vector.}

\item{verbose}{Should the function show a message with a list of changed
files?}

\item{...}{Further arguments passed to \code{utf8::transform_lines_enc()}.}
}
\description{
A wrapper for \code{\link[utf8:transform_lines_enc]{utf8::transform_lines_enc()}} which initiates the styling of
either R or Rmd files by passing the relevant transformer function for each
case.
}
