% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-code.R
\name{transform_code}
\alias{transform_code}
\title{Transform code from R, Rmd or Rnw files}
\usage{
transform_code(path, fun, ...)
}
\arguments{
\item{path}{A vector with file paths to transform.}

\item{fun}{A function that returns a character vector.}

\item{...}{Further arguments passed to \code{enc::transform_lines_enc()}.}
}
\description{
A wrapper which initiates the styling of
either R, Rmd or Rnw files by passing the relevant transformer function for each
case.
}
\keyword{internal}
