% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_file}
\alias{transform_file}
\title{Transform a file an give customized message}
\usage{
transform_file(path, fun, verbose = FALSE, ...)
}
\arguments{
\item{path}{A vector of file paths.}

\item{fun}{A function that returns a character vector.}

\item{verbose}{Should the function show a message with a list of changed
files?}

\item{...}{Further arguments passed to \code{utf8::transform_lines_enc()}.}
}
\description{
Wraps \code{utf8::transform_lines_enc()} and gives customized messages.
}
