% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_files}
\alias{transform_files}
\title{Transform files with transformer functions}
\usage{
transform_files(files, transformers)
}
\arguments{
\item{files}{A character vector with paths to the file that should be
transformed.}

\item{transformers}{A list of transformer functions that operate on parse
tables.}
}
\value{
A logical value that indicates whether or not any file was changed is
returned invisibly. If files were changed, the user is informed to
carefully inspect the changes via a message sent to the console.
}
\description{
\code{transform_files} applies transformations to file contents and writes back
the result.
}
