% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-files.R
\name{transform_files}
\alias{transform_files}
\title{Transform files with transformer functions}
\usage{
transform_files(files, transformers)
}
\arguments{
\item{files}{A character vector with paths to the file that should be
transformed.}

\item{transformers}{A list of transformer functions that operate on flat
parse tables.}
}
\description{
\code{transform_files} applies transformations to file contents and writes back
the result.
}
\section{Value}{

Invisibly returns a data frame that indicates for each file considered for
styling whether or not it was actually changed.
}

\keyword{internal}
