% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_rmd.R
\name{transform_rmd}
\alias{transform_rmd}
\title{Transform Rmd contents}
\usage{
transform_rmd(lines, transformer_fun)
}
\arguments{
\item{lines}{A character vector of lines from an Rmd file}

\item{transformer_fun}{A styler transformer function}
}
\description{
Applies the supplied transformer function to code chunks identified within
an Rmd file and recombines the resulting (styled) code chunks with the text
chunks.
}
