% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-files.R
\name{transformers_subset}
\alias{transformers_subset}
\title{Remove transformers that are not needed}
\usage{
transformers_subset(text, transformers)
}
\arguments{
\item{text}{Text to parse. Can also be the column \code{text} of the output of
\code{\link[=compute_parse_data_nested]{compute_parse_data_nested()}}, where each element is a token (instead of a
line).}

\item{transformers}{the transformers.}
}
\description{
The goal is to speed up styling by removing all rules that are only
applicable in contexts that don't occur often, e.g. for most code, we don't
expect ";" to be in it, so we don't need to apply \code{\link[=resolve_semicolon]{resolve_semicolon()}} on
every \emph{nest}.
}
\keyword{internal}
