% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{uninstall_precommit}
\alias{uninstall_precommit}
\title{Uninstall pre-commit}
\usage{
uninstall_precommit(scope = "repo", ask = "user", root = here::here())
}
\arguments{
\item{scope}{Either "repo" or "user". "repo" removes pre-commit from your
project, but you will be able to use it in other projects. With "user",
you remove the pre-commit executable in the virtual python environment
r-precommit so it won't be available in any project. When you want to do
the latter, you should first do the former.}

\item{ask}{Either "user", "repo" or "none" to determine in which case
a prompt should show up to let the user confirm his action.}

\item{root}{The path to the root directory of your project.}
}
\value{
\code{NULL} (invisibly). The function is called for its side effects.
}
\description{
Remove pre-commit from a repo or from your system.
}
\examples{
\dontrun{
uninstall_precommit()
}
}
\seealso{
Other executable managers: 
\code{\link{install_precommit}()},
\code{\link{update_precommit}()},
\code{\link{version_precommit}()}
}
\concept{executable managers}
