% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_pd.R
\name{update_indention}
\alias{update_indention}
\alias{indent_round}
\alias{indent_op}
\title{Update indention information of parse data}
\usage{
indent_round(pd, indent_by = 2)

indent_op(pd, indent_by = 2, op = op_token)
}
\arguments{
\item{pd}{A nested or flat parse table that is already enhanced with
line break and space information via \code{\link[=create_filler]{create_filler()}} or
\code{\link[=create_filler_nested]{create_filler_nested()}}.}

\item{indent_by}{How many spaces should be added after the token of interest.}
}
\description{
Update indention information of parse data
}
