% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_precommit}
\alias{update_precommit}
\title{Update the pre-commit executable}
\usage{
update_precommit()
}
\value{
The exit status of the conda update command (invisible).
}
\description{
Updates the conda installation of the upstream framework pre-commit. This
does not update the R package \code{{precommit}} and only works if you choose
conda as your installation method. If you have problems updating, we suggest
deleting the conda environment \code{r-precommit} (if you are sure nothing but
pre-commit depend on it) and do a fresh installation with
\code{\link[=install_precommit]{install_precommit()}}.
}
\seealso{
Other executable managers: 
\code{\link{install_precommit}()},
\code{\link{uninstall_precommit}()},
\code{\link{version_precommit}()}
}
\concept{executable managers}
