% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_precommit}
\alias{use_precommit}
\title{Set up pre-commit}
\usage{
use_precommit(force = FALSE, open = TRUE, path_root = here::here())
}
\arguments{
\item{force}{Whether to replace an existing config file.}

\item{open}{Whether or not to open the .pre-commit-config.yaml after
it's been placed in your repo.}
}
\description{
Set up pre-commit
}
\details{
\itemize{
\item installs pre-commit in your current directory.
\item sets up a template \code{.pre-commit-config.yaml}.
\item autoupdates the template to make sure you get the latest versions of the
hooks.
\item Open the config file if RStudio is running.
}
}
\seealso{
Other helpers: \code{\link{open_config}}
}
\concept{helpers}
