% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{use_precommit}
\alias{use_precommit}
\title{Set up pre-commit}
\usage{
use_precommit(force = FALSE, open = TRUE, path_root = here::here())
}
\arguments{
\item{force}{Whether to replace an existing config file.}

\item{open}{Whether or not to open the .pre-commit-config.yaml after
it's been placed in your repo.}

\item{path_root}{The path to the root directory of your project.}
}
\description{
Get started.
}
\section{When to call this function?}{

\itemize{
\item You want to add pre-commit support to a git repo. This involves adding
a pre-commit config file and making sure git will call the hooks before
the next commit.
\item You use a repo that has such a config file but for pre-commit to become
active, you need to make sure git knows that it should call pre-commit.
}
}

\section{What it does the funciton do?}{

\itemize{
\item installs pre-commit in your current directory.
\item sets up a template \code{.pre-commit-config.yaml}.
\item autoupdates the template to make sure you get the latest versions of the
hooks.
\item Open the config file if RStudio is running.
}
}

\seealso{
Other helpers: 
\code{\link{open_config}()}
}
\concept{helpers}
