% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsed.R
\name{verify_roundtrip}
\alias{verify_roundtrip}
\title{Verify parse data modifications}
\usage{
verify_roundtrip(pd, text)
}
\arguments{
\item{pd}{A parse table.}

\item{text}{A character vector with the initial text to compare against.}
}
\value{
If the verification is successful, \code{pd} is returned, with empty
lines at the end of \code{text} stripped. \cr
Otherwise, an error is thrown.
}
\description{
Check whether serializing the parse data results in the same
number of lines as the initial data that should be styled.
}
\seealso{
\code{\link[=serialize_parse_data_flat]{serialize_parse_data_flat()}}
}
