% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit.R
\name{visit}
\alias{visit}
\alias{pre_visit}
\alias{post_visit}
\title{Visit'em all}
\usage{
pre_visit(pd_nested, funs)

post_visit(pd_nested, funs)
}
\arguments{
\item{pd_nested}{A nested parse table.}

\item{funs}{A list of transformer functions.}
}
\description{
Apply a list of functions to each level in a nested parse table.
\code{pre_visit()} applies \code{funs} before it proceeds to the children,
(that is, starts from the outermost level of nesting progressing
to the innermost level), \code{post_visit()} proceeds to its children
before applying the functions (meaning it first applies the functions
to the innermost level of nesting first and then going outwards).
}
\seealso{
Other visitors: 
\code{\link{visit_one}()}
}
\concept{visitors}
\keyword{internal}
