% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-create.R
\name{wrap_expr_in_curly}
\alias{wrap_expr_in_curly}
\title{Wrap an expression in curly braces}
\usage{
wrap_expr_in_curly(pd, stretch_out = FALSE)
}
\arguments{
\item{pd}{A parse table.}

\item{stretch_out}{Whether or not to create a line break after the opening
curly brace and before the closing curly brace.}
}
\description{
Adds curly braces to an expression (represented as a parse table) if there
are none.
}
