% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-tokens.R
\name{wrap_subexpr_in_curly}
\alias{wrap_subexpr_in_curly}
\title{Wrap a sub-expression in curly braces}
\usage{
wrap_subexpr_in_curly(pd, ind_to_be_wrapped, indent_by, space_after)
}
\arguments{
\item{pd}{A parse table.}

\item{ind_to_be_wrapped}{The indices of the rows that should be wrapped
into a new expression.}

\item{space_after}{How many spaces should be inserted after the closing brace.}
}
\description{
Wraps some rows of a parse table into a sub-expression.
}
\keyword{internal}
